% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abdata}
\alias{abdata}
\title{Alburnoides bipunctatus species data from GBIF and iNaturalist}
\format{
A \code{tibble} 2130 rows and 3 columns.
}
\usage{
data(abdata)
}
\description{
A \code{tibble} Data from GBIF (https://www.gbif.org/) and iNaturalist (https://www.inaturalist.org/)
}
\details{
The species data was collated from the Global Biodiversity Information Facility and iNaturalist
}
\examples{

data("abdata")
abdata


}
\keyword{datasets}
\keyword{freshwater}
\keyword{information}
\keyword{platform}
