% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinranges.R
\name{basinranges}
\alias{basinranges}
\title{Records within a particular basin range based on Tesedesco at al., 2017}
\usage{
basinranges(
  occurrences,
  species,
  lat,
  lon,
  verbose = TRUE,
  discard = TRUE,
  output = "records"
)
}
\arguments{
\item{occurrences}{Species occurrences for a particular to check for the basin where they lie in}

\item{species}{species name considered as its recorded in the occurrences}

\item{lat, lon}{latitude and longitude column names in the species occurrences records.}

\item{verbose}{Whether implementation messages are shown. Default \code{TRUE}.}

\item{discard}{Whether to discard records with basins or not. Default \code{FALSE}.}

\item{output}{Whether to produce the records or the basin files. Default \code{records}.}
}
\value{
species occurrence records and their basins were they are found.
}
\description{
Records within a particular basin range based on Tesedesco at al., 2017
}
\examples{

\dontrun{

data("efidata")

sal <- subset(efidata, scientificName=='Anguilla anguilla')

br <- basinranges(occurrences = sal, species = "Anguilla anguilla",
                  lat = 'decimalLatitude', lon = 'decimalLongitude')
}


}
