% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierclassification.R
\name{batch_extract}
\alias{batch_extract}
\title{Extract oultiers for multiple species}
\usage{
batch_extract(x)
}
\arguments{
\item{x}{Species list}
}
\value{
data frame of outliers for multiple species
}
\description{
Extract oultiers for multiple species
}
