% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierclassification.R
\name{bm}
\alias{bm}
\title{Identifies the best method for outlier detection for a single species.}
\usage{
bm(
  x,
  var,
  sp = NULL,
  threshold = 0.6,
  mode = "all",
  methods = NULL,
  warn = T,
  verbose = TRUE
)
}
\arguments{
\item{x}{List of dataframes for each methods used to identify outliers in mult_detect function.}

\item{var}{variable used to identify outliers in the univariate outlier detection methods such as zcore, IQR, boxplots, and species optimal ranges.}

\item{sp}{species name or index if multiple species are considered during outlier detection.}

\item{threshold}{Maximum value to denote an absolute outlier.}

\item{mode}{the mode is 'all' for all methods or custom for selected methods.}

\item{methods}{indicate the methods to consider including jaccard, overlap,sorensen, ocindex, and smc.}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}

\item{verbose}{if TRUE then messgaes and warnings will be produced. Default FALSE}
}
\value{
best method for identifying outliers.
}
\description{
Identifies the best method for outlier detection for a single species.
}
