% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidetect.R
\name{broad_classify}
\alias{broad_classify}
\title{Outlier detection method broad classification.}
\usage{
broad_classify(category)
}
\arguments{
\item{category}{The different outlier categories including \code{mult}, \code{uni} and \code{ref}}
}
\value{
\code{vector} method broad categories
}
\description{
Outlier detection method broad classification.
}
\examples{

x <- broad_classify(category = "mult")

}
