% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{bulkopt}
\alias{bulkopt}
\title{Check for outliers for multiple species using temperature ranges from FishBase.}
\usage{
bulkopt(
  data,
  sp = NULL,
  optimal,
  species,
  var,
  min,
  max,
  output,
  sciname = NULL,
  check.names = TRUE
)
}
\arguments{
\item{data}{Dataframe to check for outliers.}

\item{sp}{If dataframe is used, then sp is the column with species names.}

\item{optimal}{Dataframe with standard species optimal ranges retrieved from FIshBase.}

\item{species}{Species column name for the standard database with optimal parameters.}

\item{var}{Environmental parameter considered in flagging suspicious outliers}

\item{min}{Minimum temperature column from the standard optimal dataframe.}

\item{max}{Maximum temperature column from the standard optimal dataframe.}

\item{output}{output Either clean: for dataframe with no suspicious outliers or outlier: to retrun dataframe with only outliers.}

\item{sciname}{Only used if only one species is considered}

\item{check.names}{Check names for the species and default is TRUE}
}
\value{
Dataframe with or with no outliers.
}
\description{
Check for outliers for multiple species using temperature ranges from FishBase.
}
\examples{

\dontrun{

#' library(terra)

#species data from online databases

data(efidata)
data(jdsdata)

multispecies <- merge_datasets(datasets = list(jds = jdsdata, efi = efidata),
                    lats = 'lat',
                    lons = 'lon',
                    species = c('speciesname','scientificName')

multspchecked <- check_names(data = multispecies, colsp='species', pct=90, merge=TRUE)

#preclean and extract

danube <- system.file('extdata/danube/basinfinal.shp', package='specleanr')

danubebasin <- sf::st_read(danube, quiet=TRUE)

#Get environmental data

#worldclim_bio <- env_download(var='bio', resolution = 10, basin = danube, folder='worlclimddata')

worldclim <- terra::rast(system.file('extdata/worldclim.tiff', quiet=TRUE, package='specleanr'))

optimal_df <- ecoranges(multspchecked, colsp = 'speciescheck', range=c('n', 'a'),
                              basin = 'Danu')

precleaned <- precleaner(data = gbchecked,
                          raster= worldclim ,
                          lat = 'decimalLatitude',
                          lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = danubebasin,
                          multiple = FALSE,
                          minpts = 10)

#only retain species with optimal ranges

finaldata <- precleaned \%>\% filter(species\%in\%optimal_df$Species)

#outliers
multioultiers <- bulkopt(data = finaldata,
                        var = 'bio1',
                        sp = 'species',#preclenaed data
                        optimal = optimal_df,
                        min = 'TempMin',
                        max = 'TempMax',
                        species = 'Species', #optimal data
                        output='outlier')

multiout_clean <- bulkopt(data = finaldata,
                        var = 'bio1',
                        sp = 'species',#preclenaed data
                        optimal = optimal_df,
                        min = 'TempMin',
                        max = 'TempMax',
                        species = 'Species', #optimal data
                        output='clean')

}



}
