% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{check_dates}
\alias{check_dates}
\title{Harmonise dates form different data sets}
\usage{
check_dates(x, month = NULL)
}
\arguments{
\item{x}{vector or list of dates}

\item{month}{position of the month, if a date of 03/04/2020. Indicate if the month is 04,
therefore month = 'middle' or month= 'lastfirst' if the month is 03}
}
\value{
Standardised dates
}
\description{
Harmonise dates form different data sets
}
\examples{

\dontrun{

data(efidata)

efidates <- check_dates(x= efiplus$Date, month ='middle')

}

}
