% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checknatura.R
\name{check_natura}
\alias{check_natura}
\title{Check whether species is on NATURA2000 list}
\usage{
check_natura(x, quiet = TRUE, details = FALSE)
}
\arguments{
\item{x}{Dataframe or vector of names.}

\item{quiet}{TRUE to hide implementation messages and \code{FALSE} to show messages. Default \code{FALSE}.}

\item{details}{TRUE to give species details such as the country where it is found or \code{FALSE} to whether the species is NATURA2000. Default \code{FALSE}.}
}
\value{
species list
}
\description{
Check whether species is on NATURA2000 list
}
\examples{

\dontrun{

salmo <- check_natura(x="Salmo trutta")

}
}
