% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precleaning.R
\name{check_outlier}
\alias{check_outlier}
\title{Determine the outliers at level during pre-cleaning}
\usage{
check_outlier(data, colsp, lat, lon, basin = NULL)
}
\arguments{
\item{data}{Data frame with species data.}

\item{colsp}{variable with species names.}

\item{lat}{variable for latitude column name.}

\item{lon}{variable for longitude column name.}

\item{basin}{variable to determine how many records discarded that are not in the basin.}
}
\value{
Records of discarded at every step of data checks.
}
\description{
Determine the outliers at level during pre-cleaning
}
\examples{

\dontrun{

library(sf)

data(jdsdata)

danube <- system.file('extdata/danube/basinfinal.shp', quiet=TRUE, package='specleanr')

danubebasin <- st_read(danube)

df <- check_outliers(data = efidata, lat ='decimalLatitude', lon = 'decimalLongitude',
 basin = 'danube', colsp = 'scientificName')

}


}
