% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_packages}
\alias{check_packages}
\title{Check for packages to install and respond to use}
\usage{
check_packages(pkgs)
}
\arguments{
\item{pkgs}{list of packages to install}
}
\value{
error message for packages to install
}
\description{
Check for packages to install and respond to use
}
