% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checks}
\alias{checks}
\title{Post checks for PCA and bootstrapping}
\usage{
checks(y, nboots, th, var)
}
\arguments{
\item{y}{list of PCA and bootstrapped output.}

\item{nboots}{Number of bootstrapping}

\item{th}{threshold for identifying absolute outlier from bootstrapped samples.}

\item{var}{variable of interest.}
}
\description{
Post checks for PCA and bootstrapping
}
