% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandataextraction.R
\name{clean_data}
\alias{clean_data}
\title{Extract final clean data using either absolute or best method generated outliers.}
\usage{
clean_data(
  data,
  outliers,
  sp = NULL,
  mode = "best",
  colsp = NULL,
  threshold = NULL,
  verbose = FALSE,
  warn = FALSE,
  pabs = 0.1,
  autothreshold = FALSE,
  loess = FALSE
)
}
\arguments{
\item{data}{\code{dataframe}. The reference data for the species used in outlier detection.}

\item{outliers}{\code{string}. Output from the outlier detection process.}

\item{sp}{\code{string}. species name. NULL if a single species is considered in outlier detection.}

\item{mode}{\code{character}. Either \code{abs} to use absolute outliers to filter data or \code{best} to outliers from best method.}

\item{colsp}{\code{string}. A parameter to be used if the \code{data} is a data frame and the user must indicate the column wih species names.}

\item{threshold}{\code{numeric}. Value to consider whether the outlier is an absolute outlier or not.}

\item{verbose}{\code{logical}. Produces messages or not. Default \strong{FALSE}.}

\item{warn}{\code{logical}. If \strong{FALSE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}

\item{pabs}{\code{numeric}. Percentage of outliers allowed to be extracted from the data. If \code{best} is used to extract outliers and the \code{pabs} is exceeded,
the absolute outliers are removed instead. This because some records  in the best methods are repeated and they will likely to remove true values as outliers.}

\item{autothreshold}{\code{vector}. Identifies the threshold with mean number of absolute outliers.The search is limited within 0.51 to 1 since thresholds less than
are deemed inappropriate for identifying absolute outliers. The autothreshold is used when \code{threshold} is set to \code{NULL}.}

\item{loess}{\code{logical}. Set to \code{TRUE} to use loess threshold optimisation to extract clean data.

###param multiple TRUE for multiple species and FALSE for single species considered during outlier detection.}
}
\value{
\code{dataframe} Cleaned dataframe of species records.
}
\description{
Extract final clean data using either absolute or best method generated outliers.
}
\seealso{
\code{\link{thresh_search}}
}
