% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandataextraction.R
\name{clean_data_2}
\alias{clean_data_2}
\title{Extract final clean data when multiple variables are considered.}
\usage{
clean_data_2(data, outliers, threshold, warn = FALSE)
}
\arguments{
\item{data}{\code{dataframe}. The reference data for the species used in outlier detection.}

\item{outliers}{\code{string}. Output from the outlier detection process.}

\item{threshold}{\code{numeric}. Value to consider whether the outlier is an absolute outlier or not.}

\item{warn}{\code{logical}. If \strong{FALSE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}
}
\value{
\code{dataframe} A dataframe with outliers replaced with NAs and is strictly used when multiple variables
         of concerns are used during outlier detection.
}
\description{
Extract final clean data when multiple variables are considered.
}
