% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandataextraction.R
\name{clean_data_extract}
\alias{clean_data_extract}
\title{Extract cleaned record for multiple species.}
\usage{
clean_data_extract(
  refdata,
  outliers,
  mode = "best",
  colsp = NULL,
  threshold = NULL,
  warn = FALSE,
  verbose = FALSE,
  autothreshold = FALSE,
  pabs = 0.1
)
}
\arguments{
\item{refdata}{Species data frame from precleaned analysis}

\item{outliers}{Output from the outlier detection}

\item{mode}{Either all to consider all similarity measure to consider suitable method or custom.
to indicate the similarity measures or interest. simple matching coefficient, jaccard, sorensen,
and overlap coefficient can be used.
Otherwise a data frame will be produced. Default FALSE.}

\item{colsp}{Column with species names.}

\item{threshold}{value to consider whether the outlier is an absolute outlier or not.}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}

\item{verbose}{Produces messages or not. Default \strong{FALSE}.}

\item{autothreshold}{Identifies the threshold with mean number of absolute outliers.The search is limited within 0.51 to 1 since thresholds less than
are deemed inappropriate for identifying absolute outliers. The autothreshold is used when \code{threshold} is set to \code{NULL}.}

\item{pabs}{Percentage of outliers allowed to be extracted from the data. If \code{best} is used to extract outliers and the \code{pabs} is exceeded,
the absolute outliers are removed instead. This because some records  in the best methods are repeated and they will likely to remove true values as outliers.}
}
\value{
Either a list or data frame of cleaned records for multiple species.
}
\description{
Extract cleaned record for multiple species.
}
\examples{

\dontrun{

data(jdsdata)

data(efidata)


matchdata <- merge_datasets(datasets = list(jds = jdsdata, efi = efidata),
                    lats = 'lat',
                    lons = 'lon',
                    species = c('speciesname','scientificName'),
                    country= c('JDS4_site_ID),
                    dates=c('sampling_date', 'Date'))

datacheck <- check_names(matchdata, colsp= 'species', pct = 90, merge =TRUE)


db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'), quiet=TRUE)


worldclim <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

rdata <- predextract(data = datacheck,
                          raster= worldclim ,
                          lat = 'decimalLatitude',
                          lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = db,
                          multiple = TRUE,
                          minpts = 10,
                          list=TRUE,
                          merge=F)


out_df <- multidetect(data = rdata, multiple = TRUE,
                       var = 'bio6',
                        output = 'outlier',
                        exclude = c('x','y'),
                        methods = c('zscore', 'adjbox',iqr', 'semiqr','hampel'),
                        kmpar =list(k = 6, method='silhouette'),
                        ifpar = list(cutoff = 0.5, size=0.7))

 #extracting use the absolute method for one species

 multextract_abs <- clean_data_extract(data = precleaned, outliers = out_df,
 var = 'bio6', type = 'absolute', threshold = 0.6, autothreshold = FALSE)

 bestmout_bm <- clean_data_extract(data = rdata, outliers = out_df,
 var = 'bio6', type = 'bestmethod', threshold = 0.6, autothreshold = FALSE)


}



}
