% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{clean_date}
\alias{clean_date}
\title{Internal function to clean dates before checking for consitencies}
\usage{
clean_date(x)
}
\arguments{
\item{x}{Date to to be cleaned}
}
\value{
cleaned date
}
\description{
Internal function to clean dates before checking for consitencies
}
\examples{

\dontrun{

date <- clean_date(x = '12 january 2010')

}


}
