% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{clean_names}
\alias{clean_names}
\title{Internal function to clean names before checking them with fishbase}
\usage{
clean_names(sp)
}
\arguments{
\item{sp}{species names}
}
\value{
corrected species
}
\description{
Internal function to clean names before checking them with fishbase
}
\examples{

\dontrun{

speciesname <- clean_names(sp='salmo trutta??')

}


}
