% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{datacleaner-class}
\alias{datacleaner-class}
\title{Outlier detection class for multiple methods}
\description{
Outlier detection class for multiple methods
}
\section{Slots}{

\describe{
\item{\code{result}}{List of data sets with outliers detected.}

\item{\code{mode}}{Either ´TRUE´ for multiple species and FALSE for one species.}

\item{\code{varused}}{The variable used for outlier detection, useful for univariate outlier detection methods.}

\item{\code{out}}{Either outliers or clean dataset outputted.}

\item{\code{methodsused}}{The methods used in outlier detection.}

\item{\code{dfname}}{the dataframe name to aid tracking it during clean data extraction.}

\item{\code{excluded}}{whether some columns were excluded during outlier detection. useful for multivariate methods where coordinates are removed from the data.}

\item{\code{pc}}{parameters for principal component analysis.}

\item{\code{bootstrap}}{parameters for bootstrapping for small data sets.}

\item{\code{nboots}}{the number of bootstraps during bootstrapping.}

\item{\code{pcvariable}}{variable to be considered during PCA.}

\item{\code{pcretained}}{the number data columns retained. the default is 3.}

\item{\code{maxrecords}}{the maximum number of records used for bootstrapping.}
}}

