% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dataonline-class}
\alias{dataonline-class}
\title{Data online class}
\description{
Data online class
}
\section{Slots}{

\describe{
\item{\code{output}}{List of datasets from each database.}
}}

