% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dateformats}
\alias{dateformats}
\title{Standard date formats}
\format{
A \code{tibble} with 2 columns and 52 rows of date formats.
}
\usage{
data(dateformats)
}
\description{
A \code{tibble} Dates standard dataframe. Extensible by the package author
}
\details{
Standard date formats for cleaning the provided dates from users
}
\examples{

\dontrun{

data("dateformats")
dateformats
}



}
\keyword{datasets}
