% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidetect.R
\name{detect}
\alias{detect}
\title{Flags outliers or no outliers for multiple outlier methods.}
\usage{
detect(
  x,
  var,
  output,
  exclude,
  optpar,
  kmpar,
  ifpar,
  lofpar,
  jkpar,
  gloshpar,
  mahalpar,
  knnpar,
  zpar,
  methods,
  verbose,
  spname,
  warn,
  missingness,
  showErrors
)
}
\arguments{
\item{x}{Dataframe with species occurrences}

\item{var}{Environmental parameter considered in flagging suspicious outliers.}

\item{output}{Either \strong{clean}; to produce a data set with no outliers, or \strong{outlier}:
to output a dataframe with outliers. Default \code{outlier}.}

\item{exclude}{Exclude variables that should not be considered in the fitting the one
class model, for example x and y columns or
latitude/longitude or any column that the user doesn't want to consider.}

\item{optpar}{Parameters for species optimal ranges like temperatures ranges \link[specleanr]{multidetect}.}

\item{kmpar}{Parameters for kmeans clustering like method and number of clusters for tuning \link[specleanr]{multidetect}.}

\item{ifpar}{Isolation forest parameter settings. Parameters of the isolation
 model that are required, include
the \strong{cutoff} to be used for denoting outliers. It ranges from \strong{0 to 1}
 but Default \strong{0.5}. Also,
the \strong{size} of data partitioning for training should be determined.
For more details check \strong{(Liu et al. 2008)}}

\item{lofpar}{Parameters for local outlier factor such as the distance matrix and mode of method implementation
such as robust and soft modes \link[specleanr]{multidetect}.}

\item{jkpar}{Parameters for reverse jack knifing mainly the mode used \link[specleanr]{multidetect}.}

\item{gloshpar}{Parameters for global local outlier score from hierarchies such as distance metric used \link[specleanr]{multidetect}.}

\item{mahalpar}{Parameters for Malahanobis distance which includes varying the mode of output \link[specleanr]{multidetect}.}

\item{knnpar}{Parameters for varying the distance matrix such euclidean or Manhattan \link[specleanr]{multidetect}.}

\item{zpar}{Parameters for z-score such mode and x parameter \link[specleanr]{multidetect}.}

\item{methods}{Outlier detection methods considered. Use \strong{\code{extractMethod}}
to get outliers implemented in the package.}

\item{verbose}{whether to return messages or not. Default FALSE.}

\item{spname}{species name being handled.}

\item{warn}{whether to return warning or not. Default TRUE.}

\item{missingness}{Allowed missing values in a column to allow a user decide whether to remove the individual columns or rows from the data sets. Default 0.1. Therefore, if
if a column has more than 10\% missing values, then it will be removed from the dataset rather than the rows.}

\item{showErrors}{show execution errors and therefore for multiple species the code will break if one of the
methods fails to execute.}
}
\value{
Dataframe with or with no outliers.
}
\description{
Flags outliers or no outliers for multiple outlier methods.
}
\references{
\enumerate{
  \item Liu FeiT, Ting KaiM, Zhou Z-H. 2008. Isolation Forest. Pages 413-422
In 2008 Eighth IEEE International Conference on Data Mining. Available from
https://ieeexplore.ieee.org/abstract/document/4781136 (accessed November 18, 2023).

}
}
\seealso{
\code{\link[specleanr]{multidetect}}
}
