% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distr_ranges}
\alias{distr_ranges}
\title{Determine species distribution ranges based on IUCN RedList.}
\usage{
distr_ranges(data, species, lat, lon, key = NULL, verbose = T)
}
\arguments{
\item{data}{\code{dataframe}. species data frame to check with International Union for Conservation of Nature.}

\item{species}{\code{string}. Species column name in the data.}

\item{lat, lon}{\code{coordinates}. Columns with latitude and longitude in the species data frame.}

\item{key}{\code{string}. API key to access IUCN RedList data. Use rl_use_iucn() to apply for it.}

\item{verbose}{\code{logical}. if \code{TRUE} processing messages or warning will be produced.}
}
\value{
\code{dataframe} Species records within the known distribution ranges of the species according to IUCN RedList database.
}
\description{
Determine species distribution ranges based on IUCN RedList.
}
\examples{

\dontrun{
df <- 1

}

}
