% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{.abspath}
\alias{.abspath}
\title{obtain absolute path for the user}
\usage{
.abspath(dir, verbose = TRUE)
}
\arguments{
\item{dir}{to user the user directory to save data for future use.}

\item{verbose}{to show messages during implementation or not. Default \code{FALSE}.}
}
\value{
absolute path
}
\description{
obtain absolute path for the user
}
