% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{.absx}
\alias{.absx}
\title{create sub folders in the absolute path folder}
\usage{
.absx(x, var)
}
\arguments{
\item{x}{is the absolute path set in .abspath}

\item{var}{is the name of the folder with specific variables.}
}
\value{
subfolder in the absolute path
}
\description{
create sub folders in the absolute path folder
}
