% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinranges.R
\name{.gbdownload}
\alias{.gbdownload}
\title{Downloads the global basin data based on Tesedesco at al., 2017}
\usage{
.gbdownload(x = "globalbasins")
}
\arguments{
\item{x}{absolute path to cache global basin files}
}
\value{
download files for global basin
}
\description{
Downloads the global basin data based on Tesedesco at al., 2017
}
