% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataacess.R
\name{.n2000list}
\alias{.n2000list}
\title{Download NATURA2000 species list}
\usage{
.n2000list(x)
}
\arguments{
\item{x}{Absolute path}
}
\value{
species list
}
\description{
Download NATURA2000 species list
}
