% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesecology.R
\name{.nearthdownload}
\alias{.nearthdownload}
\title{NatureEarth data}
\usage{
.nearthdownload(x)
}
\arguments{
\item{x}{Absolute path}
}
\value{
shapefile form Natural Earth for Administrative boundaries
}
\description{
NatureEarth data
}
