% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checknatura.R
\name{downloadnatura}
\alias{downloadnatura}
\title{Download NATURA2000 data}
\usage{
downloadnatura(x, timeout = 180)
}
\arguments{
\item{x}{path to save European data files.}

\item{timeout}{Delay time while downloading}
}
\description{
Download NATURA2000 data
}
