% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efidata}
\alias{efidata}
\title{EFIPLUS data used to develop ecological sensitivity parameters for riverine species in European streams and rivers.}
\format{
A \code{tibble} 99 rows and 23 columns.
}
\usage{
data(efidata)
}
\description{
A \code{tibble}
}
\details{
BQEs sensitivity to global/climate change in European rivers:
implications for reference conditions and pressure-impact-recovery chains (Logez et al. 2012). An extract has been made for
usage in this package but for more information write to ihg@boku.ac.at
}
\examples{

data("efidata")
efidata

}
\references{
Logez M, Belliard J, Melcher A, Kremser H, Pletterbauer F, Schmutz S, Gorges G, Delaigue O, Pont D. 2012.
Deliverable D5.1-3: BQEs sensitivity to global/climate change in European rivers: implications for reference conditions
and pressure-impact-recovery chains.
}
\keyword{European}
\keyword{dataset}
\keyword{wide}
