% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eif}
\alias{eif}
\title{Computes the empirical influence function for each values in the dataset}
\usage{
eif(x, var)
}
\arguments{
\item{x}{Outlier checked data}

\item{var}{variable of interest}
}
\description{
Computes the empirical influence function for each values in the dataset
}
