% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesecology.R
\name{endemicity}
\alias{endemicity}
\title{Determines the species ranges, whether native or alien based on Freshwater Information Platform.}
\usage{
endemicity(basin, range)
}
\arguments{
\item{basin}{Basin under consideration.}

\item{range}{Whether introduced (a) or native (n)}
}
\value{
dataframe showing species ranges
}
\description{
Determines the species ranges, whether native or alien based on Freshwater Information Platform.
}
\examples{

\dontrun{

df <- endemicity(basin ='Danu', range =c('a','n'))
}

}
