% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depindep.R
\name{evaluate}
\alias{evaluate}
\title{Model evaluation metrics}
\usage{
evaluate(data, predictions, response, model = NULL, cutoff, metrics)
}
\arguments{
\item{data}{Species environmental data generated during bootstrapping or subsampling.}

\item{predictions}{Model predictions from the \code{predict} function from stats package.}

\item{response}{The presence absence or presence only for presence-only models.}

\item{model}{Parameter to set the model of choices. Only \code{GLM} and \code{RF} are allowed.}

\item{cutoff}{Defines a threshold classify the model predictions/probabilities as presence or absent. Default is 0.5. The maximum is 1 and lowest is 0.}

\item{metrics}{Either to only consider threshold-dependent, threshold-independent or all evaluation metrics to determine model performance.}
}
\value{
evaluation metrics values
}
\description{
Model evaluation metrics
}
\seealso{
\code{\link{boot}}
}
