% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similaritytests.R
\name{ext_outliers}
\alias{ext_outliers}
\title{Extract outliers for a one species}
\usage{
ext_outliers(x, sp = NULL)
}
\arguments{
\item{x}{list of outlier outputs.}

\item{sp}{species name or index in the list from datacleaner output. NULL for a single species}
}
\value{
data frame of outliers for each method
}
\description{
Extract outliers for a one species
}
\examples{

\dontrun{

data(efidata)

db <- sf::read_sf(system.file('extdata/danube/basinfinal.shp', package = "specleanr"), quiet = TRUE)

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package = "specleanr"))

checkname <- check_names(data=efidata, colsp ='scientificName', pct = 90, merge = T)

extdf <- pred_extract(data = checkname, raster = wcd,
                      lat = 'decimalLatitude', lon = 'decimalLongitude',
                     colsp = 'speciescheck',
                     list = TRUE,verbose = F,
                     minpts = 6,merge = F)#basin removed

 #outlier detection

outliers <- multidetect(data = extdf, output='outlier',
                       exclude = c('x','y'), multiple = TRUE,
                       methods = c('mixediqr', "iqr", "kmeans", "mahal"),
                       kmpar =list(k=6, method='silhouette', mode='soft'))

extractoutlier <- ext_outliers(x=outliers, sp = 5)

}


}
