% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extentvalues}
\alias{extentvalues}
\title{To check for a bounding box}
\usage{
extentvalues(x, par = NULL)
}
\arguments{
\item{x}{raster, shapefile or list of bounding box values.}

\item{par}{indicate the database being queried to handing the issues of bounding box settings.}
}
\value{
extent values from raster, shapefile and bounding box
}
\description{
To check for a bounding box
}
