% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataextraction.R
\name{extract_clean}
\alias{extract_clean}
\title{Extract final clean data using either absolute or best method generated outliers.}
\usage{
extract_clean(
  data,
  outliers,
  sp = NULL,
  var,
  type = "bestmethod",
  threshold = 0.6,
  mode = "all",
  verbose = FALSE,
  warn = TRUE
)
}
\arguments{
\item{data}{Species data frame from precleaned analsyis}

\item{outliers}{Output from the outlier detection}

\item{sp}{species name. NULL if a single species is considered in outlier detection.}

\item{var}{variable used to identify outliers in the univariate outlier detection methods and species optimal ranges.}

\item{type}{Either absolute to use absolute outliers to filter data or bestmethod to outliers from best method}

\item{threshold}{value to consider whether the outlier is an absolute outlier or not.}

\item{mode}{Either all to consider all similarity measure to consider best method or custom
to indicate the similarity measures or interest. simple matching coefficient, jaccard, sorensen,
and overlap coefficient can be used.}

\item{verbose}{Produces messages or not. Default \strong{FALSE}.}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}
}
\value{
cleaned data frame of species records.
}
\description{
Extract final clean data using either absolute or best method generated outliers.
}
\examples{

\dontrun{

a = 1+1

}

}
