% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comp.R
\name{extract_performance}
\alias{extract_performance}
\title{Extract the model performance}
\usage{
extract_performance(modeloutput, type = "test")
}
\arguments{
\item{modeloutput}{The model output object where data is extracted.}

\item{type}{Either test or train data}
}
\value{
A data frame with performance metrics between reference and outlier cleaned dataset.
}
\description{
Extract the model performance
}
