% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidetect.R
\name{extractMethods}
\alias{extractMethods}
\title{List of outlier detection methods implemented in this package.}
\usage{
extractMethods()
}
\value{
List of methods
}
\description{
List of outlier detection methods implemented in this package.
}
\examples{
extractMethods()


}
