% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesecology.R
\name{fip_ranges}
\alias{fip_ranges}
\title{Extract basin ranges from Freshwater Information Platform.}
\usage{
fip_ranges(data, colsp = NULL, basin, range, pct = 90, verbose = F, sn = FALSE)
}
\arguments{
\item{data}{Data frame to extract basin species records which are either native, endemic or both.}

\item{colsp}{A variable in the data with species names}

\item{basin}{Basin names as indicated in the Freshwater Information Platform (Schmidt-Kloiber et al., 2019)}

\item{range}{For each basin, either indicate n to return only native, a: alien (Schmidt-Kloiber et al., 2019)}

\item{pct}{The percentage similarity to be considered to assign a relatively similar name from FishBase if the exact match is absent.
The higher the pct values, the higher the different similar are considered to replace a species name that is cheeked from FisBase.}

\item{verbose}{TRUE to return processing messages or FALSE for no messages. Default FALSE}

\item{sn}{Either \strong{FALSE} to return accepted names from FishBase rather than synonyms.Default FALSE.}
}
\value{
Species data frame with temperature ranges from FishBase
}
\description{
Extract basin ranges from Freshwater Information Platform.
}
\examples{

\dontrun{

data(jdsdata)


nativedanube <- ecoranges(x = jdsdata, colsp = 'speciesname',
range = 'n', verbose = T, basin = 'Danu')

#Species introduced in Danube

aliendf <- ecoranges(x = jdsdata, colsp = 'species', basin = 'Danu', range = 'a')

# Species check from FIP and FB

aliennativedf <- ecoranges(x = 'Salmo trutta', basin = 'Danu', range = c('n','a')

}


}
\references{
Schmidt-Kloiber, A., Bremerich, V., De Wever, A. et al. The Freshwater Information Platform:
a global online network providing data, tools and resources for science and policy support.
Hydrobiologia 838, 1-11 (2019).
}
