% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{fishbase}
\alias{fishbase}
\title{Collates synoynm data tables from FishBase.}
\usage{
fishbase(tables)
}
\arguments{
\item{tables}{Tables from FishBase}
}
\value{
Two datasets
}
\description{
Collates synoynm data tables from FishBase.
}
\examples{

\dontrun{

stdf <- fishbase(tables = 'synonym')

}

}
