% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prabs.R
\name{gen_background}
\alias{gen_background}
\title{Title}
\usage{
gen_background(
  occurences,
  raster,
  coords = c("geometry"),
  class = list(present = "P", absent = "A"),
  prop,
  set.seed = 1124
)
}
\arguments{
\item{set.seed}{}
}
\description{
Title
}
