% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{get_iucn}
\alias{get_iucn}
\title{Extract species IUCN status from IUCN RedList database.}
\usage{
get_iucn(x, key)
}
\arguments{
\item{x}{species name}

\item{key}{API key to access IUCN Red List data. Use rl_use_iucn() to apply for it.}
}
\value{
species IUCN category
}
\description{
Extract species IUCN status from IUCN RedList database.
}
