% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comp.R
\name{get_performance}
\alias{get_performance}
\title{Extract the model performance}
\usage{
get_performance(modeloutput, type = "test")
}
\arguments{
\item{modeloutput}{The model output object where data is extracted.}

\item{type}{Either test or train performance metrics.}
}
\value{
A data frame with performance metrics between reference and outlier cleaned dataset. This function takes the output from \code{\link{modelcomparison}} function.
}
\description{
Extract the model performance
}
