% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getdiff}
\alias{getdiff}
\title{get dataframe from the large dataframe.}
\usage{
getdiff(x, y, full = FALSE)
}
\arguments{
\item{x}{Small dataset}

\item{y}{Large dataset for intersection}

\item{full}{Whether the whole column names are checked or not. Default \code{FALSE} where only the first column is considered.
if FALSE; then the returned columns may be few or more if the considered column has less or more similar
rows across the two data sets.}
}
\value{
Data to extracted from large dataset.
}
\description{
get dataframe from the large dataframe.
}
\examples{

x = data.frame(id=c(1,2,3,4,5),  name=c('a','b','c', 'd','e'))

y=data.frame(id=c(1,2,3,4,7,6,5), tens=c(10,29,37,46,58, 34, 44),
                 name=c('a','b','c','d','e', 'f','g'))

}
