% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoutliers.R
\name{ggoutliers}
\alias{ggoutliers}
\title{Visualize the outliers identified by each method}
\usage{
ggoutliers(x, select = NULL, color = "purple", var_col = NULL, desc = TRUE)
}
\arguments{
\item{x}{. the datacleaner object}

\item{select}{. the species index or name for multiple species}

\item{color}{\code{string}. Color of the bars. Default is \code{grey}.}

\item{var_col}{\code{string} If the reference dataset is a \code{dataframe} class,  \code{multiple = TRUE}, and
\code{raw = FALSE}, provide the column with species or parameter names.}
}
\value{
ggplot object indicating outlier detection methods and number of outlier flagged.
}
\description{
Visualize the outliers identified by each method
}
