% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{guimerger}
\alias{guimerger}
\title{Title}
\usage{
guimerger(dir = NULL)
}
\arguments{
\item{dir}{Directory where the data files to be matched or bind are located. The main directory should be selected.}
}
\description{
Title
}
