% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierclassification.R
\name{hamming}
\alias{hamming}
\title{Title}
\usage{
hamming(x, sp = NULL, var, threshold, warn = TRUE)
}
\arguments{
\item{x}{jjj}

\item{sp}{jjj}

\item{var}{jjj}

\item{threshold}{jjj}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}
}
\description{
Title
}
