% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidetect.R
\name{handle_true_errors}
\alias{handle_true_errors}
\title{Catch errors during methods implementation.}
\usage{
handle_true_errors(
  func,
  fname = NULL,
  spname = NULL,
  verbose = FALSE,
  warn = FALSE,
  silence_true_errors = TRUE
)
}
\arguments{
\item{func}{Outlier detection function}

\item{fname}{function name for messaging or warning identification.}

\item{spname}{species name being handled}

\item{verbose}{whether to return messages or not. Default \code{FALSE}.}

\item{warn}{whether to return warning or not. Default TRUE.}

\item{silence_true_errors}{show execution errors and therefore for multiple species the code will break if one of the
methods fails to execute.}
}
\value{
Handle errors
}
\description{
Catch errors during methods implementation.
}
