% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depindep.R
\name{indep}
\alias{indep}
\title{threshold independent evaluation}
\usage{
indep(probs, observed, P, A)
}
\arguments{
\item{probs}{\code{numeric}. Model probabilities after predictions.}

\item{observed}{\code{vector}. The observed probabilities.}

\item{P}{\code{character}. for presence label. Important to indicate the presence/absence character to correctly compute the evaluation metric in \code{evaluate} function.}

\item{A}{\code{character}. for absence label. Important to indicate the presence/absence character to correctly compute the evaluation metric in \code{evaluate} function.}
}
\value{
Area Under the Curve values
}
\description{
threshold independent evaluation
}
\details{
The threshold-independent evalation parameter does not dependent on the cutoff to classify probability if a species is present or not.
In this package we incorporated the Area Under Receiver Operating Curve based on \strong{Hanley & McNeil Barbra (1982)}.
}
\references{
\enumerate{
  \item Hanley J. A, McNeil Barbra J. 1982. The meaning and use of the Area under
  a Receiver Operating Characteristic (ROC) Curve. Radiology 143:29-36.
  }
}
