% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jdsdata}
\alias{jdsdata}
\title{Joint Danube Survey Data}
\format{
A \code{tibble} 98 rows and 24 columns.
}
\usage{
data(jdsdata)
}
\description{
A \code{tibble} Data on a five year periodic data collection within the Danube River Basin.
For more information, please visit https://www.danubesurvey.org/jds4/about
}
\details{
Species ecological parameters such as ecological ranges both native and alien
}
\examples{

data("jdsdata")
jdsdata


}
\references{
https://www.danubesurvey.org/jds4/about
}
\keyword{datasets}
\keyword{freshwater}
\keyword{information}
\keyword{platform}
