% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kdat}
\alias{kdat}
\title{Sequential fences constants}
\format{
A \code{tibble} 101 rows and 2 columns.
}
\usage{
data(kdat)
}
\description{
A \code{tibble} data with k constants for sequential fences method.
}
\details{
k constants fro flagging outliers with several chnages in the fences.
}
\examples{

data("kdat")
kdat

}
\references{
Schwertman NC, de Silva R. 2007. Identifying outliers with sequential fences.
Computational Statistics and Data Analysis 51:3800–3810.
}
\keyword{constants}
\keyword{fences}
\keyword{sequential}
