% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinranges.R
\name{mbasinranges}
\alias{mbasinranges}
\title{Basin ranges for multiple species.}
\usage{
mbasinranges(
  occurrences,
  colsp = NULL,
  lat,
  lon,
  verbose = TRUE,
  discard = TRUE,
  batch = TRUE,
  output = "records"
)
}
\arguments{
\item{occurrences}{\code{dataframe}. Species occurrences for a particular to check for the basin where they lie in.}

\item{colsp}{\code{character}. Species column considered as its recorded in the occurrences.}

\item{lat, lon}{\code{coordinates}. Latitude and longitude column names in the species occurrences records.}

\item{verbose}{\code{logical}. Whether implementation messages are shown. Default \code{TRUE}.}

\item{discard}{\code{logical}. Whether to discard records with basins or not. Default \code{FALSE}.}

\item{batch}{\code{logical}. if \code{TRUE} all species records will be considered at once. Otherwise species-wise checks can
can be done. Default \code{TRUE}}

\item{output}{\code{character}. Whether to produce the records or the basin files. Default \code{records}.}
}
\value{
species occurrence records and their basins were they are found.
}
\description{
Basin ranges for multiple species.
}
\examples{

\dontrun{

data("efidata")

salmo1 <- efidata[efidata[,'scientificName'] == c("Anguilla anguilla", "Cottus gobio"), ]

br <- mbasinranges(occurrences = salmo1, lat = 'decimalLatitude', lon = 'decimalLongitude')

}
}
