% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesecology.R
\name{mdistr_ranges}
\alias{mdistr_ranges}
\title{Determine species distribution ranges based on IUCN RedList.}
\usage{
mdistr_ranges(data, colsp, lat, lon, key, mode, verbose = T)
}
\arguments{
\item{data}{species dataframe to check with International Union for Conservation of Nature.}

\item{colsp}{column with species names.}

\item{lat}{variable for latitude in the species data frame.}

\item{lon}{variable for longitude in the species data frame.}

\item{key}{API key to access IUCN RedList data. Use rl_use_iucn() to apply for it.}

\item{mode}{Either distribution only, or IUCN categories for species or both}

\item{verbose}{TRUE to reutn messages or warning during the processing of data. Default is FALSE}
}
\value{
species records within the known distribution ranges of the species according
 to IUCN RedList database for multiple species.
}
\description{
Determine species distribution ranges based on IUCN RedList.
}
\examples{

\dontrun{

mdst <- 12

}

}
