% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierclassification.R
\name{mult_abs}
\alias{mult_abs}
\title{Identifies absolute outliers for multiple species.}
\usage{
mult_abs(x, var, threshold = 0.6, props = FALSE, warn = TRUE)
}
\arguments{
\item{x}{List of data frames for each methods used to identify outliers in mult_detect function.}

\item{var}{variable used to identify outliers in the univariate outlier detection methods such as zcore, IQR, boxplots, and species optimal ranges.}

\item{threshold}{Maximum value to denote an absolute outlier.}

\item{props}{To output the proportional absoluteness for each outlier}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}
}
\value{
vector or absolute outliers, best outlier detection method or data frame of absolute outliers and their
proportions
}
\description{
Identifies absolute outliers for multiple species.
}
