% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierclassification.R
\name{mult_bm}
\alias{mult_bm}
\title{Identify best method for outlier removal for multiple species using majority votes.}
\usage{
mult_bm(
  x,
  var,
  threshold,
  mode = "all",
  methods = NULL,
  warn = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{Output from the outlier detection.}

\item{var}{variable used to identify outliers in the univariate outlier detection methods such as zcore, IQR, boxplots, and species optimal ranges.}

\item{threshold}{value to consider whether the outlier is an absolute outlier or not.}

\item{mode}{Either all to consider all similarity measure to consider best method or custom
to indicate the similarity measures or interest. simple matching coefficient, jaccard, sorensen,
and overlap coefficient can be used.}

\item{methods}{if mode is set to custom then methods should be indicated to be used to identify outliers
These include simple matching coefficient, jaccard, sorensen, and overlap coefficient can be used.}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}

\item{verbose}{Produce messages on the process or not. Default \strong{FALSE}.}
}
\value{
best method for outlier detection for each species
}
\description{
Identify best method for outlier removal for multiple species using majority votes.
}
