% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataextraction.R
\name{multextract}
\alias{multextract}
\title{Extract cleaned record for multiple species.}
\usage{
multextract(
  data,
  outliers,
  var,
  type = "bestmethod",
  colsp = NULL,
  threshold,
  mode = "all",
  list = TRUE,
  warn = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Species data frame from precleaned analysis}

\item{outliers}{Output from the outlier detection}

\item{var}{variable considered in outlier detection.}

\item{type}{Either absolute to use absolute outliers to filter data or suitable method to outliers from suitable method.}

\item{colsp}{Column with species names.}

\item{threshold}{value to consider whether the outlier is an absolute outlier or not.}

\item{mode}{Either all to consider all similarity measure to consider suitable method or custom
to indicate the similarity measures or interest. simple matching coefficient, jaccard, sorensen,
and overlap coefficient can be used.}

\item{list}{If TRUE the a list of multiple species data will be produced.
Otherwise a data frame will be produced. Default FALSE.}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}

\item{verbose}{Produces messages or not. Default \strong{FALSE}.}
}
\value{
Either a list or data frame of cleaned records for multiple species.
}
\description{
Extract cleaned record for multiple species.
}
