% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naturalist}
\alias{naturalist}
\title{Dataset with species NATURA2000 categories.}
\format{
A \code{tibble} 388303 rows and i9 columns.
}
\usage{
data(naturalist)
}
\description{
A \code{tibble} The data consist of 388303 rows and 19 columns which represents the
       different species on the list. The list was downloaded from the EEA website and manually
       archived in the package for use. For more information and citation
}
\details{
The data is extracted from \code{(https://www.eea.europa.eu/themes/biodiversity/natura-2000)}.
}
\examples{

\dontrun{

data("naturalist")
naturalist
}

}
\references{
https://www.eea.europa.eu/themes/biodiversity/natura-2000.
}
\keyword{NATURA2000}
\keyword{fish.}
