% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataacess.R
\name{naturaranges}
\alias{naturaranges}
\title{Check if the records are obtained within the NATURA2000 sites}
\usage{
naturaranges(
  x,
  species,
  lat,
  lon,
  verbose = FALSE,
  discard = TRUE,
  batch = FALSE
)
}
\arguments{
\item{x}{Dataframe input with species records. Only dataframe accepted.}

\item{species}{Column name containing species names.}

\item{lat}{Column in the data with latitude coordinates. Only decimal degrees accepted.}

\item{lon}{Column in the data with longitude coordinates. Only decimal degrees accepted.}

\item{verbose}{TRUE, to show processing messages. Default FALSE}

\item{discard}{TRUE to discard records outside the NATURA2000 sites and FALSE to retained then in the final dataset. Defualt TRUE}

\item{batch}{TRUE, if the records have more than one species one species.}
}
\value{
Dataset with records found in NATURA2000 polygon sites.
}
\description{
Check if the records are obtained within the NATURA2000 sites
}
\details{
The fucntion is aimed at identifying records that located in the NATURA2000 protected sites mostly for invasive species.
    If records of invasive species are recorded in the protected, then management measures are required to avoid the loss
    of ecosytem diversity and species which are threatened with extinction. Also, foe species distribution modelling,
    records in the protected areas are vital to predicts the areas that are occupied/invaded to fdervie prioroty based
    conservation. To better understand the species on the NATURA2000 list, \code{check_natura} fucntion may be used.
}
