% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndata}
\alias{ndata}
\title{Standard species ecological ranges}
\format{
A \code{tibble} 637 rows and 129 columns.
}
\usage{
data(ndata)
}
\description{
A \code{tibble} Extensible from Freshwater Information Platform  (Schmidt-Kloiber et al., 2019)
}
\details{
Species ecological parameters such as ecological ranges both native and alien
}
\examples{

\dontrun{

data("ndata")
ndata
}

}
\references{
Schmidt-Kloiber, A., Bremerich, V., De Wever, A. et al. The Freshwater Information Platform:
a global online network providing data, tools and resources for science and policy support. Hydrobiologia 838, 1–11 (2019).
 https://doi.org/10.1007/s10750-019-03985-5
}
\author{
Anthony Basooma <anthony.basooma@boku.ac.at>
}
\keyword{datasets}
\keyword{freshwater}
\keyword{information}
\keyword{platform}
