% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similaritytests.R
\name{overlap}
\alias{overlap}
\title{Identifies best outlier detection method using Overlap coefficient.}
\usage{
overlap(
  x,
  sp = NULL,
  var,
  threshold = NULL,
  warn = FALSE,
  autothreshold = FALSE
)
}
\arguments{
\item{x}{List of data frames for each methods used to identify outliers in mult_detect function.}

\item{sp}{species name or index if multiple species are considered during outlier detection.}

\item{var}{variable used to identify outliers in the univariate outlier detection methods such as zcore, IQR, boxplots, and species optimal ranges.}

\item{threshold}{Maximum value to denote an absolute outlier. The threshold ranges from \code{0} which indicates a point has not been flagged by any outlier detection method
as an \code{outlier} or \code{1}, when means the record is an absolute or true outlier sicen it has been identified by all methods. At both extremes, at low threshold values,
many records are classified, which may be due to individual method weakness or strength and data distribution. Also, at higher threshold values, the true outliers are retained
Fo example, if 10 methods are considered and 9 methods flags a record as an outlier, If a cut off 1 is used, then that particular record is retained.
Therefore the \code{default} cutoff is 0.6 but \code{autothreshold} can be used to select the appropriate threshold.}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}

\item{autothreshold}{Identifies the threshold with mean number of absolute outliers.The search is limited within 0.51 to 1 since thresholds less than
are deemed inappropriate for identifying absolute outliers. The autothreshold is used when \code{threshold} is set to \code{NULL}.}
}
\value{
best method for identifying outliers.
}
\description{
Identifies best outlier detection method using Overlap coefficient.
}
\examples{

\dontrun{

data(efidata)

db <- sf::read_sf(system.file('extdata/danube/basinfinal.shp', package = "specleanr"), quiet = TRUE)

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package = "specleanr"))

checkname <- check_names(data=efidata, colsp ='scientificName', pct = 90, merge = T)

extdf <- pred_extract(data = checkname, raster = wcd,
                    lat = 'decimalLatitude', lon = 'decimalLongitude',
                     colsp = 'speciescheck',
                     list = TRUE,verbose = F,
                     minpts = 6,merge = F)#basin removed

 #outlier detection

outliers <- multidetect(data = extdf, var = 'bio6', output='outlier',
                       exclude = c('x','y'), multiple = TRUE,
                       methods = c('mixediqr', "iqr", "kmeans", "mahal"),
                       kmpar =list(k=6, method='silhouette', mode='soft'))

overlapout <- overlap(x = outliers, sp= 1, threshold = 0.2, var = 'bio6')#


}


}
