% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pca}
\alias{pca}
\title{Implement principal component analysis for dimension reduction}
\usage{
pca(data, npc, q)
}
\arguments{
\item{data}{Environmental dataframe}

\item{npc}{Number of principal components to be retained. Default is 2}

\item{q}{To show the cumulative total variance explained by the \code{npc} selected.}
}
\description{
Implement principal component analysis for dimension reduction
}
