% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pcboot}
\alias{pcboot}
\title{To package both principal component analysis and bootstrapping.}
\usage{
pcboot(pb, var, pc, boot, pcvar)
}
\arguments{
\item{pb}{the principal component or bootstrapped data}

\item{var}{The variable of concern, which is vital for univariate outlier detection methods}

\item{pc}{Whether principal component analysis will be computed. Default \code{FALSE}}

\item{boot}{Whether bootstrapping will be computed. Default \code{FALSE}}

\item{pcvar}{Principal component analysis to e used for outlier detection after PCA. Default \code{PC1}}
}
\description{
To package both principal component analysis and bootstrapping.
}
