% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoutliers.R
\name{plot,datacleaner,ANY-method}
\alias{plot,datacleaner,ANY-method}
\title{Visualize the outliers identified by each method}
\usage{
\S4method{plot}{datacleaner,ANY}(x, y, raw = TRUE, colsp = NULL)
}
\arguments{
\item{x}{. the datacleaner object}

\item{y}{. the species index or name for multiple species}

\item{raw}{Whether total number of outlier flagged in each method are provided (\code{TRUE}) or
\code{FALSE} to return the percentage outlier contribution for each method.}

\item{colsp}{\code{string} If the reference dataset is a \code{dataframe} class,  \code{multiple = TRUE}, and
\code{raw = FALSE}, provide the column with species or parameter names.}
}
\value{
ggplot object indicating outlier detection methods and number of outlier flagged.
}
\description{
Visualize the outliers identified by each method
}
