% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdopt.R
\name{search_threshold}
\alias{search_threshold}
\title{Determine the threshold using Locally estimated or weighted Scatterplot Smoothing.}
\usage{
search_threshold(
  data,
  outliers,
  sp = NULL,
  plot = FALSE,
  var_col = NULL,
  warn = FALSE,
  verbose = FALSE,
  colors = c("darkblue", "orange")
)
}
\arguments{
\item{data}{\code{Dataframe}. The reference dataframe were absolute outliers will be removed.}

\item{outliers}{\code{datacleaner}. Datacleaner output with outliers flagged in \code{multidetect} function.}

\item{sp}{\code{string}. Species name or index if multiple species are
considered during outlier detection.}

\item{plot}{\code{logical}. to show plot of loess fitted function with minima and maxima (optimal threshold and clean data).}

\item{var_col}{\code{string}. A column with species names if \code{dataset} for species is a dataframe not a list.
See \code{\link{pred_extract}} for extracting environmental data.}

\item{warn}{\code{logical}. If \strong{TRUE}, warning on whether absolute
outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}

\item{verbose}{\code{logical}. If true, then messages about the outlier flagging will be displayed.}

\item{colors}{\code{vector}. Colors for both the true data and the loess fitted data lines.}
}
\value{
Returns \code{numeric} of most suitable threshold at maxima of the loess smoothing.
}
\description{
Determine the threshold using Locally estimated or weighted Scatterplot Smoothing.
}
