% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,datacleaner-method}
\alias{show,datacleaner-method}
\title{set method for displaying output details after outlier detection.}
\usage{
\S4method{show}{datacleaner}(object)
}
\arguments{
\item{object}{The data model for outlier detection.}
}
\value{
prints the datacleaner class for this package.
}
\description{
set method for displaying output details after outlier detection.
}
