% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,dataonline-method}
\alias{show,dataonline-method}
\title{Method for viewing details of the downloaded data.}
\usage{
\S4method{show}{dataonline}(object)
}
\arguments{
\item{object}{The data online model for displaying data.}
}
\description{
Method for viewing details of the downloaded data.
}
