% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show}
\alias{show}
\title{Method for viewing detials of the downloaded data.}
\usage{
show(object)

show(object)
}
\arguments{
\item{object}{The data model for outlier detection.}

\item{dataonline}{Data online class.}

\item{datacleaner}{Data cleaner class.}
}
\description{
Method for viewing detials of the downloaded data.

set method for displaying output details after outlier detection.
}
