% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierclassification.R
\name{smc}
\alias{smc}
\title{Identify best outlier detection method using simple matching coefficient.}
\usage{
smc(x, sp = NULL, var, threshold, warn = TRUE)
}
\arguments{
\item{x}{List of data frames for each methods used to identify outliers in mult_detect function.}

\item{sp}{species name or index if multiple species are considered during outlier detection.}

\item{var}{variable used to identify outliers in the univariate outlier detection methods such as zcore, IQR, boxplots, and species optimal ranges.}

\item{threshold}{Maximum value to denote an absolute outlier.}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}
}
\value{
best method for identifying outliers.
}
\description{
Identify best outlier detection method using simple matching coefficient.
}
