% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{suggested.packages}
\alias{suggested.packages}
\title{Check for suggested packages.}
\usage{
suggested.packages(
  listpkgs = c("shiny", "shinydashboard", "DT", "dplyr"),
  reason = "open the R Shiny Application",
  quiet = TRUE
)
}
\arguments{
\item{listpkgs}{A \code{list} of packages to be suggested.}

\item{reason}{A \code{string} of character to describe the reason why packages are suggested.}

\item{quiet}{\code{logical} Default TRUE, for no messages.}
}
\value{
install suggested packages
}
\description{
Check for suggested packages.
}
