% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierdetection.R
\name{tcatch}
\alias{tcatch}
\title{Title}
\usage{
tcatch(f, fname = NULL, spname = NULL, verbose = FALSE, warn = TRUE)
}
\arguments{
\item{f}{Outlier detection function}

\item{fname}{function name for messaging or warning identification.}

\item{spname}{species name being handled}

\item{verbose}{whether to return messages or not. Default FALSE.}

\item{warn}{whether to return warning or not. Default TRUE.}
}
\value{
Handle errors
}
\description{
Title
}
