% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdopt.R
\name{thresh_search}
\alias{thresh_search}
\title{Determine the threshold using Locally estimated or weighted Scatterplot Smoothing.}
\usage{
thresh_search(
  data,
  outliers,
  sp = NULL,
  plot = FALSE,
  var_col = NULL,
  warn = FALSE,
  verbose = FALSE,
  colors = c("darkblue", "orange"),
  tuneLoess = seq(0.75, 1, 0.1)
)
}
\arguments{
\item{data}{\code{dataframe}. The reference data for the species used in outlier detection.}

\item{outliers}{\code{string}. Output from the outlier detection process.}

\item{sp}{\code{string}. species name. NULL if a single species is considered in outlier detection.}

\item{plot}{\code{logical}. to show plot of loess fitted function with minima and maxima (optimal threshold and clean data).}

\item{warn}{\code{logical}. If \strong{FALSE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}

\item{verbose}{\code{logical}. Produces messages or not. Default \strong{FALSE}.}

\item{colors}{\code{vector}. Colors for both the true data and the loess fitted data lines.}

\item{tuneLoess}{\code{vector}. Values to tune the span of the loess modeling. The best is obtained when RMSE
is minimum.}
}
\value{
Returns \code{numeric} of most suitable threshold at maxima of the loess smoothing.
}
\description{
Determine the threshold using Locally estimated or weighted Scatterplot Smoothing.
}
