% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{troutdata}
\alias{troutdata}
\title{Salmo trutta fario species data from GBIF and iNaturalist}
\format{
A \code{tibble} 1100 rows and 7 columns.
}
\usage{
data(troutdata)
}
\description{
A \code{tibble} Data from GBIF (https://www.gbif.org/) and iNaturalist (https://www.inaturalist.org/)
}
\details{
The species data was collated from the Global Biodiversity Information Facility and iNaturalist
}
\examples{

\dontrun{

data("troutdata")
troutdata
}


}
\keyword{datasets}
\keyword{freshwater}
\keyword{information}
\keyword{platform}
