% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ttdata}
\alias{ttdata}
\title{Thymallus thymallus species data from GBIF and iNaturalist}
\format{
A \code{tibble} 100 rows and 8 columns.
}
\usage{
data(ttdata)
}
\description{
A \code{tibble} Data from GBIF (https://www.gbif.org/) and iNaturalist (https://www.inaturalist.org/)
}
\details{
The species data was collated from the Global Biodiversity Information Facility and iNaturalist
}
\examples{

data("ttdata")
ttdata


}
\keyword{datasets}
\keyword{freshwater}
\keyword{information}
\keyword{platform}
