% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliermethods.R
\name{xkmedoid}
\alias{xkmedoid}
\title{Title}
\usage{
xkmedoid(data, k = 2, metric = "manhattan", output, exclude = NULL, x = 1.5)
}
\arguments{
\item{data}{the environmental data where outliers are examined from.}

\item{k}{the number of cluster centers to form cluster around it. Since kmedoid using the raw valaues from thr dataset, it is not insesetive to outliers.}

\item{metric}{Different distance based matrics including the Euclidean and Mahattan are implemented.}

\item{output}{Either clean or outliers dataset. Defualt \code{outlier} to output outliers dataset.}

\item{exclude}{columns to remove in implementtimg kmedoid algorithms, for example, the cordinates. This becuase kmediod is a multivariate algorithm which use all the data.}

\item{x}{a constant to detemrine outliers.}
}
\value{
clean or outlier data set after outlier detection
}
\description{
Title
}
