% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliermethods.R
\name{adjustboxplots}
\alias{adjustboxplots}
\title{Adjust the boxplots bounding fences using medcouple to flag suspicious outliers (Hubert & Vandervieren 2008).}
\usage{
adjustboxplots(data, var, output, a = -4, b = 3, coef = 1.5)
}
\arguments{
\item{data}{Dataframe to check for outliers}

\item{var}{Environmental parameter considered in flagging suspicious outliers}

\item{output}{Either clean: for dataframe with no suspicious outliers or outlier: to return dataframe with only outliers}

\item{a}{Constant for adjusted boxplots}

\item{b}{Constant for adjusted boxplots}

\item{coef}{Constant for adjusted boxplots}
}
\value{
Dataframe with or with no outliers.
}
\description{
Adjust the boxplots bounding fences using medcouple to flag suspicious outliers (Hubert & Vandervieren 2008).
}
\examples{

\dontrun{

data("efidata")

gbd <- check_names(data = efidata, colsp='scientificName', pct=90, merge=TRUE)

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'))

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

refdata <- pred_extract(data = gbd, raster= wcd , lat = 'decimalLatitude', lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = db,
                          multiple = FALSE,
                          minpts = 10)

 adout <- adjustboxplots(data = refdata[['Salmo trutta']], var = 'bio6', output='outlier')

}

}
\references{
Hubert M, Vandervieren E. 2008. An adjusted boxplot for skewed distributions.
Computational Statistics and Data Analysis 52:5186–5201.
}
\author{
Anthony Basooma (anthony.basooma@boku.ac.at)
}
