% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similaritytests.R
\name{batch_extract}
\alias{batch_extract}
\title{Extract oultiers for multiple species}
\usage{
batch_extract(x)
}
\arguments{
\item{x}{Species list}
}
\value{
data frame of outliers for multiple species
}
\description{
Extract oultiers for multiple species
}
\examples{

\dontrun{

#mult data extract

data(efidata)

db <- sf::read_sf(system.file('extdata/danube/basinfinal.shp', package = "specleanr"), quiet = TRUE)

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package = "specleanr"))

checkname <- check_names(data=efidata, colsp ='scientificName', pct = 90, merge = T)

extdf <- pred_extract(data = checkname, raster = wcd,
                      lat = 'decimalLatitude', lon = 'decimalLongitude',
                     colsp = 'speciescheck',
                     list = TRUE,verbose = F,
                     minpts = 6,merge = F)#basin removed

 #outlier detection

outliers <- multidetect(data = extdf, var = 'bio6', output='outlier',
                       exclude = c('x','y'), multiple = TRUE,
                       methods = c('mixediqr', "iqr", "kmeans", "mahal"),
                       kmpar =list(k=6, method='silhouette', mode='soft'))

extractbatch <- batch_extract(x=outliers)
}

}
