% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similaritytests.R
\name{batch_extract}
\alias{batch_extract}
\title{Extract outliers for multiple species}
\usage{
batch_extract(x)
}
\arguments{
\item{x}{Species list}
}
\value{
\code{dataframe} Outliers for multiple species.
}
\description{
Extract outliers for multiple species
}
\examples{

\dontrun{

#mult data extract

data(efidata)

db <- sf::read_sf(system.file('extdata/danube/basinfinal.shp', package = "specleanr"), quiet = TRUE)

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package = "specleanr"))

checkname <- check_names(data=efidata, colsp ='scientificName', pct = 90, merge = T)

extdf <- pred_extract(data = checkname, raster = wcd,
                      lat = 'decimalLatitude', lon = 'decimalLongitude',
                     colsp = 'speciescheck',
                     list = TRUE,verbose = F,
                     minpts = 6,merge = F)#basin removed

 #outlier detection

outliersdf <- multidetect(data = extdf, output='outlier', var = 'bio6',
                         exclude = c('x','y'), multiple = TRUE,
                         methods = c('mixediqr', "iqr", "mahal", "iqr", "logboxplot"),
                         showErrors = FALSE, warn = TRUE, verbose = FALSE, sdm = TRUE)

extractbatch <- batch_extract(x=outliersdf)
}

}
