% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{boots}
\alias{boots}
\title{To implement bootstrapping procedures. Sampling with replacement.}
\usage{
boots(data, boots, seed, pca)
}
\arguments{
\item{data}{Environmental data}

\item{boots}{Number of bootstraps}

\item{seed}{Random seed to ensure reproduciblity}

\item{pca}{Whether bootstrapping is conducted on data after principal component analysis.}
}
\description{
To implement bootstrapping procedures. Sampling with replacement.
}
