% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{boots}
\alias{boots}
\title{Creating sub-samples for modeling evaluation.}
\usage{
boots(data, nboots, setseed = 1234, testprob = 0.2)
}
\arguments{
\item{data}{Data to be sub sampled and partitioned.}

\item{nboots}{the number of iterations of sub sampling to be conducted.}

\item{setseed}{For reproducibility of the sub-samples across platforms and users.}

\item{testprob}{The probability to be used for partitioning data.}
}
\value{
A list of training and test data sets for use in the sdmfit.
}
\description{
Creating sub-samples for modeling evaluation.
}
\examples{

\dontrun{

data(jdsdata)

#select species with enough records

dataprep <- envextract(occurences = jdsdata, raster = worldclim,
                      lat = "lat", lon = "lon", binary = FALSE, prop = 0.8)

nboots <- boots(data = dataprep, nboots = 10, testprob = 0.3)

}
}
\seealso{
\code{\link{sdmfit}}
}
\author{
##
}
