% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{check.exclude}
\alias{check.exclude}
\title{indicate excluded columns.}
\usage{
check.exclude(x, exclude, quiet = TRUE)
}
\arguments{
\item{x}{dataframe with columns to exclude.}

\item{exclude}{columns to exclude.}

\item{quiet}{TRUE if implementation messages to be shown. Default \code{FALSE}.}
}
\value{
columns excluded from the dataset.
}
\description{
indicate excluded columns.
}
