% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataacess.R
\name{check_natura}
\alias{check_natura}
\title{Check whether species is on NATURA2000 list}
\usage{
check_natura(x, colsp)
}
\arguments{
\item{x}{Absolute path}

\item{colsp}{column name with species names}
}
\value{
species list
}
\description{
Check whether species is on NATURA2000 list
}
