% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariatemethods.R
\name{distboxplot}
\alias{distboxplot}
\title{Title}
\usage{
distboxplot(data, var, output, p1 = 0.025, p2 = 0.975)
}
\arguments{
\item{data}{hh}

\item{var}{hh}

\item{output}{hh}

\item{p1}{hh}

\item{p2}{hhh}
}
\value{
h hh
}
\description{
Title
}
\examples{
\dontrun{
data("efidata")

gbd <- check_names(data = efidata, colsp='scientificName', pct=90, merge=TRUE)

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'))

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

refdata <- pred_extract(data = gbd, raster= wcd , lat = 'decimalLatitude', lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = db,
                          multiple = FALSE,
                          minpts = 10)

 bxout <- distboxplot(data = refdata[['Salmo trutta']], var = 'bio6', output='outlier')
}
}
