% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distr_ranges}
\alias{distr_ranges}
\title{Determine species distribution ranges based on IUCN RedList.}
\usage{
distr_ranges(data, species, lat, lon, key = NULL, verbose = T)
}
\arguments{
\item{data}{species data frame to check with International Union for Conservation of Nature.}

\item{species}{species name}

\item{lat}{variable for latitude in the species data frame.}

\item{lon}{variable for longitude in the species data frame.}

\item{key}{API key to access IUCN RedList data. Use rl_use_iucn() to apply for it.}

\item{verbose}{if TRUE processing messages or warning will be produced.}
}
\value{
species records within the known distribution ranges of the species according to IUCN RedList database.
}
\description{
Determine species distribution ranges based on IUCN RedList.
}
\examples{

\dontrun{
df <- 1

}

}
