% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliermethods.R
\name{ecological_ranges}
\alias{ecological_ranges}
\title{Check for outliers for multiple species using temperature ranges from FishBase.}
\usage{
ecological_ranges(
  data,
  var = NULL,
  output = "outlier",
  species = NULL,
  optimumSettings = list(optdf = NULL, optspcol = NULL, mincol = NULL, maxcol = NULL,
    ecoparam = NULL, direction = NULL),
  minval = NULL,
  maxval = NULL,
  lat = NULL,
  lon = NULL,
  ecoparam = NULL,
  direction = NULL,
  pct = 80,
  checkfishbase = FALSE,
  mode = "temp",
  warn = TRUE
)
}
\arguments{
\item{data}{Dataframe to check for outliers.}

\item{var}{Environmental parameter considered in flagging suspicious outliers}

\item{output}{output Either clean: for dataframe with no suspicious outliers or outlier: to retrun dataframe with only outliers.}

\item{species}{Species column name for the standard database with optimal parameters.}

\item{optimumSettings}{o}

\item{minval}{Minimum temperature column from the standard optimal dataframe.}

\item{maxval}{Maximum temperature column from the standard optimal dataframe.}

\item{lat}{o}

\item{lon}{o}

\item{ecoparam}{o}

\item{direction}{o}

\item{pct}{o}

\item{checkfishbase}{o}

\item{mode}{o}

\item{warn}{o}
}
\value{
Dataframe with or with no outliers.
}
\description{
Check for outliers for multiple species using temperature ranges from FishBase.
}
\examples{

\dontrun{

library(terra)

#species data from online databases

#data(efidata)
#data(jdsdata)

#multispecies <- merge_datasets(datasets = list(jds = jdsdata, efi = efidata),
                    #lats = 'lat',
                   # lons = 'lon',
                    #species = c('speciesname','scientificName')

#multspchecked <- check_names(data = multispecies, colsp='species', pct=90, merge=TRUE)

#preclean and extract

#danube <- system.file('extdata/danube/basinfinal.shp', package='specleanr')

#danubebasin <- sf::st_read(danube, quiet=TRUE)

#Get environmental data

#worldclim_bio <- env_download(var='bio', resolution = 10, basin = danube, folder='worlclimddata')

#worldclim <- terra::rast(system.file('extdata/worldclim.tiff', quiet=TRUE, package='specleanr'))

#optimal_df <- ecoranges(multspchecked, colsp = 'speciescheck', range=c('n', 'a'),
                              basin = 'Danu')

#precleaned <- precleaner(data = gbchecked,
                          #raster= worldclim ,
                          #lat = 'decimalLatitude',
                          #lon= 'decimalLongitude',
                          #colsp = 'speciescheck',
                          #basin = danubebasin,
                          #multiple = FALSE,
                          #minpts = 10)

#only retain species with optimal ranges

#finaldata <- precleaned \%>\% filter(species\%in\%optimal_df$Species)

#outliers
#multioultiers <- ecological_ranges(data = finaldata,
                        #var = 'bio1',
                        #sp = 'species',#preclenaed data
                        #optimal = optimal_df,
                        #min = 'TempMin',
                        #max = 'TempMax',
                        #species = 'Species', #optimal data
                        #output='outlier')

#multiout_clean <- ecological_ranges(data = finaldata,
                        #var = 'bio1',
                        #sp = 'species',#preclenaed data
                        #optimal = optimal_df,
                        #min = 'TempMin',
                        #max = 'TempMax',
                        #species = 'Species', #optimal data
                        #output='clean')

}

}
