% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmdataprep.R
\name{envextract}
\alias{envextract}
\title{Generating pseudo absences from raster layers for distribution modeling.}
\usage{
envextract(
  occurences,
  raster,
  lat = NULL,
  lon = NULL,
  geom = FALSE,
  binary = FALSE,
  labels = NULL,
  prop = 0.8,
  set.seed = 1124,
  positive = NULL,
  missingness = 0.1,
  exclude = NULL,
  vifcutoff = 0.9,
  verbose = FALSE
)
}
\arguments{
\item{occurences}{Species occurrences data with coordinates or geometry to enable generate pseudo absences and extract environmental data from the points.}

\item{raster}{A raster layer with species environmental data to be used during modelling. Different sources, include WORLDCLIM}

\item{lat, lon}{If the species occurrences don't have the geometry column or not spatial vector file, the latitude and longitude must be provided for data extraction form the raster layers.}

\item{geom}{Is used in data extraction when the species occurrences geometry column instead of latitude and longitude.}

\item{binary}{Either \code{FALSE} if the species dataset do not have label column for presence absence. Therefore, the \code{FALSE} is used
when the parameter \code{label} is NULL. If only species presences and absences are indicated, the user should select \code{TRUE} option.}

\item{labels}{If the dataset has presence absence column, it should be indicated here to be used for extracting the labels.
Character labels for example allowed, for example P for presences and A for absences recommended.}

\item{prop}{The proportion of pseudo absences to presences. Default of 1 is used. Therefore equal number of pseudo absences are generated commensurate to the species.}

\item{set.seed}{to ensure reproduciblity the seed is set. Same psudoabsences will be produced during each run.}

\item{positive}{if \code{TRUE} is used, the user should indicate the positive label. For example, P for presence label. This is important in fitting the
models and computing the evaluation metrics.}

\item{missingness}{Allowed missing values in a column to allow a user decide whether to remove the individual columns or rows from the data sets. Default 0.1. Therefore, if
if a column has more than 10\% missing values, then it will be removed from the dataset rather than the rows.}

\item{exclude}{Remove non numeric variables from the data.}

\item{vifcutoff}{Used in assessing multicolinearity in environmental predictors using correlation from vifcor function from usdm package \code{Naimi et al., 2014}.}

\item{verbose}{Either \code{TRUE} to return messages or \code{FALSE} for no implementation messages. Default \code{FALSE}}
}
\value{
Labeled presence absence data.
}
\description{
Generating pseudo absences from raster layers for distribution modeling.
}
\details{
In this package, this fucntion was developed to aid in preparing the data for developing species distribution models in \code{\link{sdmfit}}. If only species occurrences
or absences are available, the function generates pseudo absences required in species distribution modeling. If the data is \code{binary} with both \strong{presences and absences},
 the user has to set which label denotes the presence of species using the \code{positive} parameter. The user should indicate the column where the labels are found.
  To account for multicolinearity among the environmental predictors, we used the \code{vifcor} function from usdm to drop highly correlated environmental
  predictors. Therefore, user can adjust the \code{vifcutoff} to range from \code{0 to 1}.
}
\examples{

\dontrun{

data(efidata)

worldclim <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

#select species with enough records
#1. Presence data: the final dataset has both presence and pseudo absences
 dfprep <- envextract(occurences = efidata, raster = worldclim,
                        lat = "decimalLatitude", lon = "decimalLongitude",
                        binary = FALSE, prop = 0.8)

#2. Presence absence

efidata$pa <- rep(c("A", "P"), c(nrow(efidata)/2, nrow(efidata)/2))

dfprep <- envextract(occurences = efidata, raster = worldclim, lat = "decimalLatitude",
                    lon = "decimalLongitude", binary = TRUE,
                    positive = "P", labels = "pa", vifcutoff = 0.9)

 #3. With sf geometry column

 efisfdata <- sf::st_as_sf(efidata, coords = c("decimalLongitude",
 "decimalLatitude"), crs= sf::st_crs(4326))

 dfprep_geom <- envextract(occurences = efisfdata, raster = worldclim,
                          geom = 'geometry', binary = TRUE,
                          positive = "P", labels = "pa", vifcutoff = 0.9)
}

}
\references{
\enumerate{
  \item Naimi, B., Hamm, N. A., Groen, T. A., Skidmore, A. K., & Toxopeus, A. G. (2014).
  Where is positional uncertainty a problem for species distribution modelling?. Ecography, 37(2), 191-203.
}
}
