% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierclassification.R
\name{ext_outliers}
\alias{ext_outliers}
\title{Extract outliers for a one species}
\usage{
ext_outliers(x, sp = NULL)
}
\arguments{
\item{x}{list of outlier outputs.}

\item{sp}{species name or index in the list from datacleaner output. NULL for a single species}
}
\value{
data frame of outliers for each method
}
\description{
Extract outliers for a one species
}
\examples{

\dontrun{

#out_df <- multi_detect()

#df <- get_all(out)

}


}
