% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{extract_online}
\alias{extract_online}
\title{Extracts data from lists of files obtained from online databases.}
\usage{
extract_online(online, verbose = FALSE)
}
\arguments{
\item{online}{List of online data sets, Only dataonline class accepted.}

\item{verbose}{To show extraction messages. Default \code{FALSE}}
}
\value{
Extracts and arranges data from the online databases.
}
\description{
Extracts data from lists of files obtained from online databases.
}
\examples{

\dontrun{

gbdata <- getdata(data= 'Gymnocephalus baloni', gbiflim = 100, inatlim = 100,vertlim = 100)

extractdata <- extract_online(online = gbdata)

}



}
