% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishdata}
\alias{fishdata}
\title{Fish species occurrences obtained from GBIF, iNaturalist and VertNet.}
\format{
A \code{tibble} 400 rows and 8 columns.
}
\usage{
data(fishdata)
}
\description{
A \code{tibble}
}
\details{
The data was incorporated for testing the package only and the licence for citation and further use are included in the file.
}
\examples{

data("fishdata")
fishdata

}
\references{
gbif, iNaturalist and VertNet
}
\keyword{Danube}
\keyword{data.}
\keyword{fish}
\keyword{species}
