% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesecology.R
\name{geo_ranges}
\alias{geo_ranges}
\title{Checks for geographic ranges from FishBase}
\usage{
geo_ranges(
  x,
  colsp = NULL,
  lat,
  lon,
  verbose = F,
  pct = 90,
  sn = FALSE,
  output = "clean",
  warn = FALSE
)
}
\arguments{
\item{x}{Dataframe with species records to check for geographic ranges from FishBase.}

\item{colsp}{Column with species names from the data set.}

\item{lat}{Column with latitude names and decimal latitudes are accepted.}

\item{lon}{Column with longitude names and decimal longitudes are accepted.}

\item{verbose}{TRUE and messages will show. Default FALSE:}

\item{pct}{The percentage similarity of species names during standardization from FishBase.}

\item{sn}{TRUE and synonyms will be generated and not accepted ones. Default is FALSE, where species accepted names will be produced.}

\item{output}{clean to return geographical ranges.}

\item{warn}{FALSE, not to generate warnings and TRUE for warnings. Default is FALSE:}
}
\value{
Dataframe with geographical corrected ranges for species from FishBase.
}
\description{
Checks for geographic ranges from FishBase
}
