% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Download species records from online database.}
\usage{
getdata(
  data,
  colsp = NULL,
  bbox = NULL,
  isFish = TRUE,
  gbiflim = 1e+06,
  vertlim = 1000,
  inatlim = 3000,
  verbose = FALSE,
  warn = FALSE,
  ...
)
}
\arguments{
\item{data}{\code{dataframe}, \code{list}, \code{vector}, \code{string}. data to retrieve records from online
databases (GBIF, VertNET, and iNaturalist).}

\item{colsp}{\code{string}. A variable of species names. Provided if data is a data frame, so not
required for lists and vector.}

\item{bbox}{\code{vector} or \code{sf}. Bounding box to limit the download of records within a particular area. Otherwise all
records from the GBIF will be downloaded. These can be provided in two forms,
either a shapefile \code{(sf)} class accepted or provide a list of
\code{xmin}, \code{ymin}, \code{xmax}, and \code{ymax}.}

\item{isFish}{\code{logical}. To indicate if the occurrence records extracted are for fish taxa or not. This allows to clean the species names
accordingly. For other taxa a different name checks is conducted. Default is \code{TRUE}.}

\item{gbiflim}{\code{integer}. Limits on the records from the Global Biodiversity Information Platform}

\item{vertlim}{\code{integer}. Limits on the records from VertNET.}

\item{inatlim}{\code{integer}. Limits on the records from iNaturalist database.}

\item{verbose}{\code{logical}. \strong{TRUE} if detailed messages should be indicated and \strong{FALSE}
if download messages are not needed. Default \strong{TRUE}.}

\item{warn}{\code{logical}. To indicate if warning messages should be shown. Default \code{FALSE}.}

\item{...}{More function for species data download can be used.
See \code{rgbif::occ_data} for more information, \code{rinat::get_inat_obs}, and
\code{rvertnet::searchbyterm}.}
}
\value{
Lists of species records from online databases
}
\description{
Download species records from online database.
}
\details{
Note always check the validity of the species name with standard database FishBase or World Register of Marine Species.
       If the records are more than 50000 in GBIF, and extent can be provide to limit the download.
}
\examples{

\dontrun{

gbaloni <- check_names(data = 'gymnocephalus baloni', pct=90)

gbdata <- getdata(data=gbaloni, gbiflim = 100, inatlim = 100, vertlim = 100)

#Get for two species
sp_records <- getdata(data=c('Gymnocephalus baloni', 'Hucho hucho'),gbiflim = 100,
                            inatlim = 100,
                            vertlim = 100)

#3. Generate species list from a dataframe

data(efidata)

danube <- system.file('extdata/danube/basinfinal.shp', package='specleanr')

danubebasin <- sf::st_read(danube, quiet=TRUE)

datachecked <- check_names(data= efidata,colsp = 'species', pct=90, merge = TRUE)#clean names

Onlinedf <- getdata(data = datachecked, colsp='speciescheck', bbox=danubebasin)

}


}
