% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggenvironmentalspace.R
\name{ggenvironmentalspace}
\alias{ggenvironmentalspace}
\title{Title}
\usage{
ggenvironmentalspace(
  qcdata,
  xvar,
  yvar,
  xlab = NULL,
  ylab = NULL,
  ncol = 3,
  nrow = 3,
  setcolors = NULL,
  colornames = NULL,
  legendposition = "default",
  pointsize = 1,
  themebackground = "bw"
)
}
\arguments{
\item{qcdata}{zzz}

\item{xvar}{zzz}

\item{yvar}{zz}
}
\description{
Title
}
