% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoutliers.R
\name{ggoutliers}
\alias{ggoutliers}
\title{Visualize the outliers identified by each method}
\usage{
ggoutliers(x, select = NULL, color = "purple", desc = TRUE)
}
\arguments{
\item{x}{. the datacleaner object}

\item{select}{\code{vect}. Enter selected groups to be displayed especially if they are greater than 10.
For example if the species are more than 10, the plot will be done in batches.}

\item{color}{\code{string}. Color of the bars. Default is \code{grey}.}

\item{desc}{\code{logical} To either arrange the bars in ascending or descending order.}
}
\value{
ggplot object indicating outlier detection methods and number of outlier flagged.
}
\description{
Visualize the outliers identified by each method
}
