% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotoutliers.R
\name{ggoutliers}
\alias{ggoutliers}
\title{Visualize the outliers identified by each method}
\usage{
ggoutliers(x, y, raw = TRUE, color = "purple")
}
\arguments{
\item{x}{. the datacleaner object}

\item{y}{. the species index or name for multiple species}

\item{raw}{Whether total number of outlier flagged in each method are provided (\code{TRUE}) or
\code{FALSE} to return the percentage outlier contribution for each method.}

\item{color}{\code{string}. Color of the bars. Default is \code{grey}.}
}
\value{
ggplot object indicating outlier detection methods and number of outlier flagged.
}
\description{
Visualize the outliers identified by each method
}
