% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggperform.R
\name{ggperform}
\alias{ggperform}
\title{Plot to the performance metrics across the thresholds used in model comparison}
\usage{
ggperform(modelout, eval = c("auc", "Accuracy"), type = "test", cutoff = 0.5)
}
\arguments{
\item{modelout}{The model comparison output obtained.}

\item{eval}{The model evaluation metrics. Check \code{\link{indep}} and \code{\link{dep}} for details.}

\item{type}{Either test data output \code{test} or training data output \code{train}. Default \code{test}.}

\item{cutoff}{The value to limit the thresholds to display. Default \code{0.5}}
}
\value{
ggplot2 output with performance metrics such as accuracy, sensitivity, specificity..
}
\description{
Plot to the performance metrics across the thresholds used in model comparison
}
