% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggperform.R
\name{ggperform}
\alias{ggperform}
\title{Plot to the performance metrics across the thresholds used in model comparison}
\usage{
ggperform(
  modelout,
  eval = c("auc", "Accuracy"),
  type = "test",
  scales = "free",
  FUN = "mean",
  scolor = c("grey75", "grey50")
)
}
\arguments{
\item{modelout}{\code{list}. The model comparison output obtained from the \code{modelcomaprison} function.}

\item{eval}{\code{vector}. The model evaluation metrics. Check \code{\link{indep}} and \code{\link{dep}} for details.}

\item{type}{\code{string}. Either test data output \code{test} or training data output \code{train}. Default \code{test}.}

\item{scales}{\code{string}. options to for facets scales based on ggplot2 object. Allowed \code{free_y, free_x, free}.}

\item{FUN}{a function to summarize the performance metrics. Default \code{mean}.}

\item{scolor}{\code{vector}. Two color options for Cleaned and Reference categories.}
}
\value{
ggplot2 output with performance metrics such as accuracy, sensitivity, specificity..
}
\description{
Plot to the performance metrics across the thresholds used in model comparison
}
