% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliermethods.R
\name{hampel}
\alias{hampel}
\title{Flag suspicious outliers based on the Hampel filter method..}
\usage{
hampel(data, var = NULL, output, x = 3)
}
\arguments{
\item{data}{Data frame to check for outliers}

\item{var}{Environmental parameter considered in flagging suspicious outliers}

\item{output}{Either clean: for dataframe with no suspicious outliers or outlier: to retrun dataframe with only outliers}

\item{x}{A constant to create a fence or boundary to detect outliers.}
}
\value{
Data frame with or with no outliers.
}
\description{
Flag suspicious outliers based on the Hampel filter method..
}
\details{
The Hampel filter method is a robust decision-based filter that considers
the median and MAD. Outliers lies beyond \deqn{[x-* λ*MAD; x+ λ*MAD]} and
λ of 3 was considered (Pearson et al. 2016).
}
\examples{

\dontrun{

data("efidata")

gbd <- check_names(data = efidata, colsp='scientificName', pct=90, merge=TRUE)

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'), quiet = TRUE)

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

refdata <- pred_extract(data = gbd, raster= wcd , lat = 'decimalLatitude', lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          bbox = db,
                          multiple = TRUE,
                          minpts = 10)

 hampout <- hampel(data = refdata[['Salmo trutta']], var = 'bio6', output='outlier')

}

}
\references{
Pearson Ronald, Neuvo Y, Astola J, Gabbouj M. 2016. The Class of Generalized Hampel Filters.
2546-2550 2015 23rd European Signal Processing Conference (EUSIPCO).
}
