% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jdsdata}
\alias{jdsdata}
\title{Joint Danube Survey Data}
\format{
A \code{tibble} 98 rows and 24 columns.

A \code{tibble} 1100 rows and 7 columns.
}
\usage{
data(jdsdata)

data(troutdata)
}
\description{
A \code{tibble} Data on a five year periodic data collection within the Danube River Basin.
For more information, please visit https://www.danubesurvey.org/jds4/about

A \code{tibble} Data from GBIF (https://www.gbif.org/) and iNaturalist (https://www.inaturalist.org/)
}
\details{
Species ecological parameters such as ecological ranges both native and alien

The species data was collated from the Global Biodiversity Information Facility and iNaturalist
}
\examples{

\dontrun{

data("jdsdata")
jdsdata
}





\dontrun{

data("troutdata")
troutdata
}




}
\references{
https://www.danubesurvey.org/jds4/about

https://www.danubesurvey.org/jds4/about
}
\keyword{datasets}
\keyword{freshwater}
\keyword{information}
\keyword{platform}
