% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{jknife}
\alias{jknife}
\title{Ideintfies outliers using Reverse Jackknifing method based on Chapman et al., (2005).}
\usage{
jknife(data, var, output, mode = "soft")
}
\arguments{
\item{data}{Data frame to check for outliers}

\item{var}{Environmental parameter considered in flagging suspicious outliers.}

\item{output}{Either clean: for data frame with no suspicious outliers or outlier: to return data frame with only outliers}

\item{mode}{Either robust, if a robust mode is used which uses median instead of mean and median absolute deviation from median
or mad instead of standard deviation}
}
\value{
Data frame with or with no outliers.
}
\description{
Ideintfies outliers using Reverse Jackknifing method based on Chapman et al., (2005).
}
\examples{

\dontrun{

data("efidata")

gbd <- check_names(data = efidata, colsp='scientificName', pct=90, merge=TRUE)

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'))

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

refdata <- pred_extract(data = gbd, raster= wcd , lat = 'decimalLatitude', lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = db,
                          multiple = FALSE,
                          minpts = 10)

 jkout <- jknife(data = refdata[['Salmo trutta']], var = 'bio6', output='outlier')

}


}
