% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliermethods.R
\name{mahal}
\alias{mahal}
\title{Flags outliers based on Mahalanobis distance matrix for all records.}
\usage{
mahal(data, exclude = NULL, output = "outlier", mode = "soft", pdf = 0.95)
}
\arguments{
\item{data}{Dataframe to check for outliers}

\item{exclude}{Exclude variables that should not be considered in the fitting the one class model, for example x and y columns or
latitude/longitude or any column that the user doesn't want to consider.}

\item{output}{Either clean: for a data set with no outliers or outlier: to output a data frame with outliers.}

\item{mode}{Either robust, if a robust mode is used which uses mcd() instead of mean. Default mode is soft.}

\item{pdf}{Chisqure probability distribution used for flagging outliers \code{Leys et al. 2018}.}
}
\value{
Either clean or outliers dataset
}
\description{
Flags outliers based on Mahalanobis distance matrix for all records.
}
\examples{

\dontrun{

data("efidata")

gbd <- check_names(data = efidata, colsp='scientificName', pct=90, merge=TRUE)

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'), quiet = TRUE)

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

refdata <- pred_extract(data = gbd, raster= wcd ,
                       lat = 'decimalLatitude',
                       lon= 'decimalLongitude',
                       colsp = 'speciescheck',
                      bbox = db,
                       multiple = TRUE,
                       minpts = 10)

#outliers
suppressWarnings(mahal(data = refdata1[['Salmo trutta']], exclude = c("x", "y"),
                        output='outlier'))


}
}
\references{
Leys C, Klein O, Dominicy Y, Ley C. 2018. Detecting multivariate outliers:
Use a robust variant of the Mahalanobis distance. Journal of Experimental
Social Psychology 74:150-156.
}
