% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariancemethods.R
\name{mahal}
\alias{mahal}
\title{Flags outliers based on Mahalanobis distance matrix for all records.}
\usage{
mahal(data, exclude, output, mode)
}
\arguments{
\item{data}{Dataframe to check for outliers}

\item{exclude}{Exclude variables that should not be considered in the fitting the one class model, for example x and y columns or
latitude/longitude or any column that the user doesn't want to consider.}

\item{output}{Either clean: for a data set with no outliers or outlier: to output a data frame with outliers.}

\item{mode}{Either robust, if a robust mode is used which uses mcd() instead of mean. Default mode is soft.}
}
\value{
Either clean or oultiers data set
}
\description{
Flags outliers based on Mahalanobis distance matrix for all records.
}
\examples{

\dontrun{
#' library(terra)

#species data from online databases

gbdata <- df_retronline(data='Gymnocephalus baloni', gbiflim = 100, inatlim = 100, vertlim = 100)

gbfinal <- merge_all(online = gbdata)

gbchecked <- check_names(data = gbfinal, colsp='species', pct=90, merge=TRUE)

#preclean and extract

danube <- system.file('extdata/danube/basinfinal.shp', quiet=TRUE, package='specleanr')

danubebasin <- sf::st_read(danube)

#Get environmental data

#worldclim_bio <- env_download(var='bio', resolution = 10, basin = danube, folder='worldclimdata')

worldclim <- terra::rast(system.file('extdata/worldclim.tiff', quiet=TRUE, package='specleanr'))

precleaned <- precleaner(data = gbchecked,
                          raster= worldclim ,
                          lat = 'decimalLatitude',
                          lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = danubebasin,
                          multiple = FALSE,
                          minpts = 10)

#outliers
mahal_outliers <- mahal(data = precleaned,
                         exclude = c('x','y'),
                         output='outlier')

#clean data
mahal_clean <- mahal(data = precleaned,
                         exclude = c('x','y'),
                         output='clean')


}


}
