% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{match.argc}
\alias{match.argc}
\title{Customised match function}
\usage{
match.argc(x, choices, quiet = TRUE)
}
\arguments{
\item{x}{The category with words to match}

\item{choices}{The different options or choices in a particular category that are allowed.}

\item{quiet}{Default \code{FALSE} not to return messages.}
}
\value{
choices
}
\description{
Customised match function
}
