% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinranges.R
\name{mbasinranges}
\alias{mbasinranges}
\title{Basin ranges for multiple species.}
\usage{
mbasinranges(
  occurrences,
  colsp = NULL,
  lat,
  lon,
  verbose = TRUE,
  discard = TRUE,
  batch = TRUE,
  output = "records"
)
}
\arguments{
\item{occurrences}{Species occurrences for a particular to check for the basin where they lie in}

\item{colsp}{species column considered as its recorded in the occurrences.}

\item{lat, lon}{latitude and longitude column names in the species occurrences records.}

\item{verbose}{Whether implementation messages are shown. Default \code{TRUE}.}

\item{discard}{Whether to discard records with basins or not. Default \code{FALSE}.}

\item{batch}{if \code{TRUE} all species records will be considered at once. Otherwise species-wise checks can
can be done. Default \code{TRUE}}

\item{output}{Whether to produce the records or the basin files. Default \code{records}.}
}
\value{
species occurrence records and their basins were they are found.
}
\description{
Basin ranges for multiple species.
}
\examples{

\dontrun{

data("efidata")

salmo1 <- efidata[efidata[,'scientificName'] == c("Anguilla anguilla", "Cottus gobio"), ]

br <- mbasinranges(occurrences = salmo1, lat = 'decimalLatitude', lon = 'decimalLongitude')

}
}
