% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mdistr_ranges}
\alias{mdistr_ranges}
\title{Determine species distribution ranges based on IUCN RedList.}
\usage{
mdistr_ranges(data, colsp, lat, lon, key, mode, verbose = T)
}
\arguments{
\item{data}{\code{dataframe}. species data frame to check with International Union for Conservation of Nature.}

\item{colsp}{\code{string}. The column names with species data.}

\item{lat, lon}{\code{coordinates}. Columns with latitude and longitude in the species data frame.}

\item{key}{\code{string}. API key to access IUCN RedList data. Use rl_use_iucn() to apply for it.}

\item{mode}{\code{string}. indicating whether the distribution range \code{dist} of the species should be collated or
species IUCN status \code{iucn} or both.}

\item{verbose}{\code{logical}. TRUE to return messages or warning during the processing of data. Default is FALSE}
}
\value{
species records within the known distribution ranges of the species according
 to IUCN RedList database for multiple species.
}
\description{
Determine species distribution ranges based on IUCN RedList.
}
\examples{

\dontrun{

mdst <- 12

}

}
