% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliermethods.R
\name{medianrule}
\alias{medianrule}
\title{Title}
\usage{
medianrule(data, var, output, x = 2.3)
}
\arguments{
\item{data}{hh}

\item{var}{uu}

\item{output}{kk}

\item{x}{kk}
}
\value{
kk
}
\description{
Title
}
\examples{

\dontrun{

data("efidata")

gbd <- check_names(data = efidata, colsp='scientificName', pct=90, merge=TRUE)

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'))

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

refdata <- pred_extract(data = gbd, raster= wcd , lat = 'decimalLatitude', lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = db,
                          multiple = FALSE,
                          minpts = 10)

 medout <- medianrule(data = refdata[['Salmo trutta']], var = 'bio6', output='outlier')
}

}
