% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariatemethods.R
\name{medianrule}
\alias{medianrule}
\title{Title}
\usage{
medianrule(data, var, output, x = 2.3)
}
\arguments{
\item{data}{hh}

\item{var}{uu}

\item{output}{kk}

\item{x}{kk}
}
\value{
kk
}
\description{
Title
}
