% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariatemethods.R
\name{mixediqr}
\alias{mixediqr}
\title{Title}
\usage{
mixediqr(data, var, output, x = 3)
}
\arguments{
\item{data}{kk}

\item{var}{kk}

\item{output}{jj}

\item{x}{jj}
}
\value{
kkk
}
\description{
Title
}
\examples{

\dontrun{

data("efidata")

gbd <- check_names(data = efidata, colsp='scientificName', pct=90, merge=TRUE)

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'))

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

refdata <- pred_extract(data = gbd, raster= wcd , lat = 'decimalLatitude', lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = db,
                          multiple = FALSE,
                          minpts = 10)

 logout <- mixediqr(data = refdata[['Salmo trutta']], var = 'bio6', output='outlier')
}

}
