% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comp.R
\name{modelperform}
\alias{modelperform}
\title{Title}
\usage{
modelperform(
  data,
  outliers,
  models = "GLM",
  threshold = 0.6,
  species,
  var,
  exclude = NULL,
  colsp = NULL,
  raster = worldclimfinal,
  mode = c("Reference", "threshold60"),
  nboots = 10,
  testprop = 0.2,
  coords = c("x", "y"),
  prbackground = 1,
  positive = "P"
)
}
\arguments{
\item{positive}{}
}
\description{
Title
}
