% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mth}
\alias{mth}
\title{mth datasets with constant at each confidence interval levels.}
\format{
A \code{tibble} 7 rows and 9 columns.
}
\usage{
data(mth)
}
\description{
A \code{tibble} The data consist the
}
\details{
The data is extracted from \code{(Schwertman & de Silva 2007)}.
}
\examples{

\dontrun{

data("mth")
mth
}

}
\references{
Schwertman NC, de Silva R. 2007. Identifying outliers with sequential fences.
Computational Statistics and Data Analysis 51:3800–3810.
}
\author{
Anthony Basooma <anthony.basooma@boku.ac.at>
}
\keyword{Stanadard}
\keyword{a}
\keyword{at}
\keyword{compute}
\keyword{constants}
\keyword{dataset}
\keyword{outlier}
\keyword{particular}
\keyword{thresholds.}
\keyword{to}
\keyword{used}
\keyword{with}
