% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similaritytests.R
\name{multibestmethod}
\alias{multibestmethod}
\title{Identify best method for outlier removal for multiple species using majority votes.}
\usage{
multibestmethod(
  x,
  threshold = NULL,
  warn = FALSE,
  verbose = FALSE,
  autothreshold = FALSE
)
}
\arguments{
\item{x}{Output from the outlier detection.}

\item{threshold}{value to consider whether the outlier is an absolute outlier or not.}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}

\item{verbose}{Produce messages on the process or not. Default \strong{FALSE}.}

\item{autothreshold}{Identifies the threshold with mean number of absolute outliers.The search is limited within 0.51 to 1 since thresholds less than
are deemed inappropriate for identifying absolute outliers. The autothreshold is used when \code{threshold} is set to \code{NULL}.}
}
\value{
best method for outlier detection for each species
}
\description{
Identify best method for outlier removal for multiple species using majority votes.
}
\examples{

\dontrun{
data(efidata)

db <- sf::read_sf(system.file('extdata/danube/basinfinal.shp', package = "specleanr"), quiet = TRUE)

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package = "specleanr"))

checkname <- check_names(data=efidata, colsp ='scientificName', pct = 90, merge = T)

extdf <- pred_extract(data = checkname, raster = wcd,
                      lat = 'decimalLatitude', lon = 'decimalLongitude',
                     colsp = 'speciescheck',
                     list = TRUE,verbose = F,
                     minpts = 6,merge = F)#basin removed

 #outlier detection

outliers <- multidetect(data = extdf, var = 'bio6', output='outlier',
                       exclude = c('x','y'), multiple = TRUE,
                       methods = c('mixediqr', "iqr", "kmeans", "mahal"),
                       kmpar =list(k=6, method='silhouette', mode='soft'))

multbm <- mult_bm(x = outliers, threshold = 0.2, var = 'bio6')#

}

}
