% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierdetection.R
\name{multidetect}
\alias{multidetect}
\title{Outlier detection with multiple functions.}
\usage{
multidetect(
  data,
  var,
  output,
  exclude,
  multiple,
  colsp = NULL,
  optpar = list(df = NULL, min = NULL, max = NULL, sp = NULL, colsp = NULL, sciname =
    NULL),
  kmpar = list(k = 6, method = "silhouette", mode = "soft"),
  kmedoidpar = list(k = 2, metric = "manhattan"),
  ifpar = list(cutoff = 0.5, size = 0.7),
  mahalpar = list(mode = "soft"),
  jkpar = list(mode = "soft"),
  zpar = list(type = "mild", mode = "soft"),
  gloshpar = list(k = 3, metric = "manhattan", mode = "soft"),
  knnpar = list(metric = "manhattan", mode = "soft"),
  lofpar = list(metric = "manhattan", mode = "soft", minPts = 10),
  methods,
  verbose = FALSE,
  spname = NULL,
  warn = TRUE
)
}
\arguments{
\item{data}{Dataframe or list of data sets for multiple or single species after after of extraction of environment predictors.}

\item{var}{A variable to check for outliers especially the one with directly affects species distribution such as
maximum temperature of the coldest month for bioclimatic variables ((IUCN Standards and Petitions Committee, 2022)) or
stream power index for hydromorphological parameters (Logez et al., 2012).}

\item{output}{zzzzzz}

\item{exclude}{Exclude variables that should not be considered in the fitting the one class model, for example x and y columns or
latitude/longitude or any column that the user doesn't want to consider.}

\item{multiple}{zz}

\item{colsp}{zzz}

\item{optpar}{Parameters for species optimal ranges like temperatures ranges \link[specleanr]{multidetect}.}

\item{kmpar}{Parameters for kmeans clustering like method and number of clusters for tunning \link[specleanr]{multidetect}.}

\item{kmedoidpar}{Parameters for adjusting the distance metrics used in kmedoid method \link[specleanr]{multidetect}.}

\item{ifpar}{Isolation forest parameter settings. Parameters of the isolation model that are required include
the \strong{cutoff} to be used for denoting outliers. It ranges from \strong{0 to 1} but Default \strong{0.5}. Also,
the \strong{size} of data partitioning for training should be determined. For more details check \strong{(Liu et al. 2008)}}

\item{mahalpar}{Parameters for Malahanobis distance which includes varying the mode of output \link[specleanr]{multidetect}.}

\item{jkpar}{Parameters for reverse jack knifing mainly the mode used \link[specleanr]{multidetect}.}

\item{zpar}{Parameters for z-score such mode and x parameter \link[specleanr]{multidetect}.}

\item{gloshpar}{Parameters for global local outlier score from hierarchies such as distance metric used \link[specleanr]{multidetect}.}

\item{knnpar}{Parameters for varying the distance matrix such euclidean or Manhattan \link[specleanr]{multidetect}.}

\item{lofpar}{Parameters for local outlier factor such as the distance matrix and mode of method implementation
such as robust and soft modes \link[specleanr]{multidetect}.}

\item{methods}{Outlier detection methods considered. Use \strong{\code{extractMethod}} to get outliers implemented in the package.}

\item{verbose}{whether to return messages or not. Default FALSE.}

\item{spname}{species name being handled}

\item{warn}{whether to return warning or not. Default TRUE.}
}
\description{
Outlier detection with multiple functions.
}
