% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierdetection.R
\name{multidetect}
\alias{multidetect}
\title{Allows ensemble multiple outlier detection methods.}
\usage{
multidetect(
  data,
  var,
  output = "outlier",
  exclude = NULL,
  multiple,
  colsp = NULL,
  optpar = list(optdf = NULL, ecoparam = NULL, optspcol = NULL, direction = NULL, maxcol
    = NULL, mincol = NULL, maxval = NULL, minval = NULL, checkfishbase = FALSE, mode =
    "geo", lat = NULL, lon = NULL, pct = 80, warn = FALSE),
  kmpar = list(k = 6, method = "silhouette", mode = "soft"),
  kmedoidpar = list(k = 2, metric = "manhattan", full = FALSE),
  ifpar = list(cutoff = 0.5, size = 0.7),
  mahalpar = list(mode = "soft"),
  jkpar = list(mode = "soft"),
  zpar = list(type = "mild", mode = "soft"),
  gloshpar = list(k = 3, metric = "manhattan", mode = "soft"),
  knnpar = list(metric = "manhattan", mode = "soft"),
  lofpar = list(metric = "manhattan", mode = "soft", minPts = 10),
  methods,
  verbose = FALSE,
  spname = NULL,
  warn = TRUE,
  missingness = 0.1,
  showErrors = TRUE
)
}
\arguments{
\item{data}{Dataframe or list of data sets for multiple or single species after after of extraction of environment predictors.}

\item{var}{A variable to check for outliers especially the one with directly affects species distribution such as
maximum temperature of the coldest month for bioclimatic variables \code{(IUCN Standards and Petitions Committee, 2022))} or
stream power index for hydromorphological parameters \code{(Logez et al., 2012)}.}

\item{output}{Either \strong{clean}: for a data set with no outliers, or \strong{outlier}: to output a dataframe with outliers. Default \code{outlier}.}

\item{exclude}{Exclude variables that should not be considered in the fitting the one class model, for example x and y columns or
latitude/longitude or any column that the user doesn't want to consider.}

\item{multiple}{If the multiple species considered, then multiple set to TRUE: If FALSE then multiple should be set to FALSE.}

\item{colsp}{A column with species columns if the data set for species is a dataframe not a list. see \link[specleanr]{pred_extract} for extracting environmental data.}

\item{optpar}{Parameters for species optimal ranges like temperatures ranges.}

\item{kmpar}{Parameters for k-means clustering like method and number of clusters for tuning.}

\item{kmedoidpar}{Parameters for adjusting the distance metrics used in kmedoid method.}

\item{ifpar}{Isolation forest parameter settings. Parameters of the isolation model that are required include
the \strong{cutoff} to be used for denoting outliers. It ranges from \strong{0 to 1} but Default \strong{0.5}. Also,
the \strong{size} of data partitioning for training should be determined. For more details check \strong{(Liu et al. 2008)}}

\item{mahalpar}{Parameters for Malahanobis distance which includes varying the mode of output.}

\item{jkpar}{Parameters for reverse jack knifing mainly the mode used.}

\item{zpar}{Parameters for z-score such mode and x parameter.}

\item{gloshpar}{Parameters for global local outlier score from hierarchies such as distance metric used..}

\item{knnpar}{Parameters for varying the distance matrix such as \code{Euclidean} or \code{Manhattan distance}.}

\item{lofpar}{Parameters for local outlier factor such as the distance matrix and mode of method implementation
such as robust and soft modes..}

\item{methods}{Outlier detection methods considered. Use \strong{\code{extractMethod}} to get outliers implemented in the package.}

\item{verbose}{whether to return messages or not. Default FALSE.}

\item{spname}{species name being handled.}

\item{warn}{whether to return warning or not. Default TRUE.}

\item{missingness}{Allowed missing values in a column to allow a user decide whether to remove the individual columns or rows from the data sets. Default 0.1. Therefore, if
if a column has more than 10\% missing values, then it will be removed from the dataset rather than the rows.}

\item{showErrors}{show execution errors and therefore for multiple species the code will break if one of the
methods fails to execute.}
}
\value{
Outliers or clean dataset.
}
\description{
Allows ensemble multiple outlier detection methods.
}
\details{
This function computes different outlier detection methods including univariate, multivariate and species
     ecological ranges to enables seammless comaprision and similarities in the outliers detected by each
     method. This can be done for multiple species or a single species in a dataframe or lists or dataframes
     and thereafter the outliers can be extracted using the \link[specleanr]{clean_data_extract} function.
}
\examples{

\dontrun{

data("efidata")
data("jdsdata")

matchdata <- match_datasets(datasets = list(jds = jdsdata, efi=efidata),
                            lats = 'lat',
                            lons = 'lon',
                            species = c('speciesname','scientificName'),
                            date = c('Date', 'sampling_date'),
                            country = c('JDS4_site_ID'))


datacheck <- check_names(matchdata, colsp= 'species', pct = 90, merge =TRUE)


db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'), quiet=TRUE)


worldclim <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

rdata <- pred_extract(data = datacheck,
                     raster= worldclim ,
                     lat = 'decimalLatitude',
                    lon= 'decimalLongitude',
                    colsp = 'speciescheck',
                    bbox = db,
                     multiple = TRUE,
                    minpts = 10,
                    list=TRUE,
                    merge=F)


out_df <- multidetect(data = rdata, multiple = TRUE,
                     var = 'bio6',
                     output = 'outlier',
                     exclude = c('x','y'),
                     methods = c('zscore', 'adjbox','iqr', 'semiqr','hampel', 'kmeans',
                                'logboxplot', 'lof','iforest', 'mahal', 'seqfences'))


optimal rnages in the multidetect: made up

optdata <- data.frame(species= c("Salmo trutta", "Abramis brama"),
                      mintemp = c(6, 1.6),maxtemp = c(20, 21),
                       meantemp = c(8.5, 10.4), #ecoparam
                      direction = c('greater', 'greater'))
#species record

salmoabramis <- rdata[c("Salmo trutta")]

#even if one species, please indicate multiple to TRUE, since its picked from pred_extract function

out_df <- multidetect(data = salmoabramis, multiple = TRUE,
                      var = 'bio1',
                      output = 'outlier',
                      exclude = c('x','y'),
                      methods = c('zscore', 'adjbox','iqr', 'semiqr','hampel', 'kmeans',
                                  'logboxplot', 'lof','iforest', 'mahal', 'seqfences', 'optimal'),
                      optpar = list(optdf=optdata, optspcol = 'species',
                                    mincol = "mintemp", maxcol = "maxtemp"))
#plot the number of outliers

ggoutliers(out_df, 1)

}

}
\references{
\enumerate{
  \item IUCN Standards and Petitions Committee. (2022). THE IUCN RED LIST OF THREATENED SPECIESTM Guidelines for Using the IUCN Red List
Categories and Criteria Prepared by the Standards and Petitions Committee of the IUCN Species Survival Commission.
https://www.iucnredlist.org/documents/RedListGuidelines.pdf.
\item Liu FeiT, Ting KaiM, Zhou Z-H. 2008. Isolation Forest. Pages 413–422 In 2008 Eighth IEEE International Conference on Data Mining.
Available from https://ieeexplore.ieee.org/abstract/document/4781136 (accessed November 18, 2023).
}
}
