% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidetect.R
\name{multidetect}
\alias{multidetect}
\title{Ensemble multiple outlier detection methods.}
\usage{
multidetect(
  data,
  var,
  output = "outlier",
  exclude = NULL,
  multiple,
  colsp = NULL,
  optpar = list(optdf = NULL, ecoparam = NULL, optspcol = NULL, direction = NULL, maxcol
    = NULL, mincol = NULL, maxval = NULL, minval = NULL, checkfishbase = FALSE, mode =
    "geo", lat = NULL, lon = NULL, pct = 80, warn = FALSE),
  kmpar = list(k = 6, method = "silhouette", mode = "soft"),
  ifpar = list(cutoff = 0.5, size = 0.7),
  mahalpar = list(mode = "soft"),
  jkpar = list(mode = "soft"),
  zpar = list(type = "mild", mode = "soft"),
  gloshpar = list(k = 3, metric = "manhattan", mode = "soft"),
  knnpar = list(metric = "manhattan", mode = "soft"),
  lofpar = list(metric = "manhattan", mode = "soft", minPts = 10),
  methods,
  verbose = FALSE,
  spname = NULL,
  warn = FALSE,
  missingness = 0.1,
  showErrors = TRUE,
  sdm = TRUE,
  na.inform = FALSE
)
}
\arguments{
\item{data}{\code{dataframe or list}. Data sets for multiple or single species after of extraction of environment predictors.}

\item{var}{\code{character}. A variable to check for outliers especially the one with directly affects species distribution such as
maximum temperature of the coldest month for bioclimatic variables \code{(IUCN Standards and Petitions Committee, 2022))} or
stream power index for hydromorphological parameters \code{(Logez et al., 2012)}. This parameter is
necessary for the univariate outlier detection methods such as Z-score.}

\item{output}{\code{character}. Either \code{clean}: for a data set with no outliers, or \code{outlier}: to output a dataframe with outliers. Default \code{outlier}.}

\item{exclude}{\code{vector}. Exclude variables that should not be considered in the fitting the one class model, for example \code{x} and \code{y} columns or
latitude/longitude or any column that the user doesn't want to consider.}

\item{multiple}{\code{logical}. If the multiple species are considered, then multiple must be set to \code{TRUE} and \code{FALSE} for single species.}

\item{colsp}{\code{string}. A column with species names if \code{dataset} for species is a dataframe not a list. See \code{\link{pred_extract}} for extracting environmental data.}

\item{optpar}{\code{list}. Parameters for species optimal ranges like temperatures ranges. For details check \code{\link{ecological_ranges}}.}

\item{kmpar}{\code{list}. Parameters for k-means clustering like method and number of clusters for tuning. For details, check \code{\link{xkmeans}}.}

\item{ifpar}{\code{list}. Isolation forest parameter settings. Parameters of the isolation model that are required include
the \strong{cutoff} to be used for denoting outliers. It ranges from \code{0 to 1} but Default \code{0.5}. Also,
the \strong{size} of data partitioning for training should be determined. For more details check \code{(Liu et al. 2008)}}

\item{mahalpar}{\code{list}. Parameters for Malahanobis distance which includes varying the mode of output  \code{\link{mahal}}.}

\item{jkpar}{\code{list}. Parameters for reverse jack knifing mainly the mode used. For details \code{\link{jknife}}.}

\item{zpar}{\code{list}. Parameters for z-score such as \code{mode} and \code{x} parameter. For details \code{\link{zscore}}}

\item{gloshpar}{\code{list}. Parameters for global local outlier score from hierarchies such as distance metric used. For details \code{\link{xglosh}}.}

\item{knnpar}{\code{list}. Parameters for varying the distance matrix such as \code{Euclidean} or \code{Manhattan distance}. For details \code{\link{xknn}}}

\item{lofpar}{\code{list}. Parameters for local outlier factor such as the distance matrix and mode of method implementation
such as robust and soft mode. For details \code{\link{xlof}}.}

\item{methods}{\code{vector}. Outlier detection methods considered. Use \code{\link{extractMethods}} to get outlier detection methods implemented in this package.}

\item{verbose}{\code{logical}. whether to return messages or not. Default \code{FALSE}.}

\item{spname}{\code{string}. species name being handled.}

\item{warn}{\code{logical}. Whether to return warning or not. Default \code{TRUE}.}

\item{missingness}{\code{numeric}. Allowed missing values in a column to allow a user decide whether to remove the individual
columns or rows from the data sets. Default 0.1. Therefore,
     if a column has more than 10\% missing values, then it will be removed from the dataset rather than the rows.}

\item{showErrors}{\code{logical}. Show execution errors and therefore for multiple species the code will break if one of the
methods fails to execute.}

\item{sdm}{{logical} If the user sets \code{TRUE}, strict data checks will be done including removing all non-numeric
columns from the datasets before identification of outliers. If set to \code{FALSE} non numeric columns will be left
in the data but the variable of concern will checked if its numeric. Also, only univariate methods are allowed. Check
\code{\link{broad_classify}} for the broad categories of the methods allowed.}

\item{na.inform}{\code{logical} Inform on the NAs removed in executing general datasets. Default \code{FALSE}.}
}
\value{
A \code{list} of outliers or clean dataset of \code{datacleaner} class. The different attributes are
associated with the \code{datacleaner} class from \code{multidetect} function.

\itemize{
        \item{\code{result}: \code{dataframe}. list of dataframes with the outliers flagged by each method.
        }
       \item{\code{mode}: \code{logical}. Indicating whether it was multiple TRUE or FALSE.}
        \item{\code{varused}: \code{character}. Indicating the variable used for the univariate outlier detection methods. }
        \item{\code{out}: \code{character}. Whether outliers where indicated by the user or no outlier data. }
        \item{\code{methodsused}: \code{vector}. The different methods used the outlier detection process.}
        \item{\code{dfname}: \code{character}. The dataset name for the species records.}
        \item{\code{exclude}: \code{vector}. The columns which were excluded during outlier detection if any.}
        }
}
\description{
The function allows to ensemble multiple outlier detection methods to ably compare the outliers flagged
by each method.
}
\details{
This function computes different outlier detection methods including univariate, multivariate and species
     ecological ranges to enables seamless comparison and similarities in the outliers detected by each
     method. This can be done for multiple species or a single species in a dataframe or lists or dataframes
     and thereafter the outliers can be extracted using the \code{\link{clean_data_extract}} function.
}
\examples{

\dontrun{

data("efidata")
data("jdsdata")

matchdata <- match_datasets(datasets = list(jds = jdsdata, efi=efidata),
                            lats = 'lat',
                            lons = 'lon',
                            species = c('speciesname','scientificName'),
                            date = c('Date', 'sampling_date'),
                            country = c('JDS4_site_ID'))


datacheck <- check_names(matchdata, colsp= 'species', pct = 90, merge =TRUE)


db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'), quiet=TRUE)


worldclim <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

rdata <- pred_extract(data = datacheck,
                     raster= worldclim ,
                     lat = 'decimalLatitude',
                    lon= 'decimalLongitude',
                    colsp = 'speciescheck',
                    bbox = db,
                     multiple = TRUE,
                    minpts = 10,
                    list=TRUE,
                    merge=F)


out_df <- multidetect(data = rdata, multiple = TRUE,
                     var = 'bio6',
                     output = 'outlier',
                     exclude = c('x','y'),
                     methods = c('zscore', 'adjbox','iqr', 'semiqr','hampel', 'kmeans',
                                'logboxplot', 'lof','iforest', 'mahal', 'seqfences'))


#optimal ranges in the multidetect: made up

optdata <- data.frame(species= c("Salmo trutta", "Abramis brama"),
                      mintemp = c(6, 1.6),maxtemp = c(20, 21),
                       meantemp = c(8.5, 10.4), #ecoparam
                      direction = c('greater', 'greater'))
#species record

salmoabramis <- rdata["Salmo trutta"]

#even if one species, please indicate multiple to TRUE, since its picked from pred_extract function

out_df <- multidetect(data = salmoabramis, multiple = TRUE,
                      var = 'bio1',
                      output = 'outlier',
                      exclude = c('x','y'),
                      methods = c('zscore', 'adjbox','iqr', 'semiqr','hampel', 'kmeans',
                                  'logboxplot', 'lof','iforest', 'mahal', 'seqfences', 'optimal'),
                      optpar = list(optdf=optdata, optspcol = 'species',
                                    mincol = "mintemp", maxcol = "maxtemp"))
#plot the number of outliers

#ggoutliers(out_df, 1)

}

}
\references{
\enumerate{
  \item IUCN Standards and Petitions Committee. (2022). THE IUCN RED LIST OF THREATENED SPECIESTM Guidelines for Using the IUCN Red List
Categories and Criteria Prepared by the Standards and Petitions Committee of the IUCN Species Survival Commission.
https://www.iucnredlist.org/documents/RedListGuidelines.pdf.
\item Liu, F. T., Ting, K. M., & Zhou, Z. H. (2008, December). Isolation forest.
In 2008 eighth ieee international conference on data mining (pp. 413-422). IEEE.
}
}
