% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierclassification.R
\name{ocindex}
\alias{ocindex}
\title{Identifies absolute outliers and their proportions for a single species}
\usage{
ocindex(
  x,
  sp = NULL,
  var,
  threshold = 0.6,
  absolute = FALSE,
  props = FALSE,
  warn = TRUE
)
}
\arguments{
\item{x}{List of dataframes for each methods used to identify outliers in mult_detect function.}

\item{sp}{species name or index if multiple species are considered during outlier detection.}

\item{var}{variable used to identify outliers in the univariate outlier detection methods such as zcore, IQR, boxplots, and species optimal ranges.}

\item{threshold}{Maximum value to denote an absolute outlier.}

\item{absolute}{To output absolute outliers for a species.}

\item{props}{To output the proportional absoluteness for each outlier.}

\item{warn}{If \strong{TRUE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}
}
\value{
vector or absolute outliers, best outlier detection method or data frame of absolute outliers and their
proportions
}
\description{
Identifies absolute outliers and their proportions for a single species
}
\examples{

\dontrun{
#kkk

}


}
