% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdopt.R
\name{optimal_threshold}
\alias{optimal_threshold}
\title{Optimize threshold for clean data extraction.}
\usage{
optimal_threshold(
  refdata,
  outliers,
  var_col = NULL,
  warn = FALSE,
  verbose = FALSE,
  plot = FALSE
)
}
\arguments{
\item{refdata}{\code{dataframe}. Species data frame from precleaned analysis.}

\item{outliers}{\code{string}. Output from the outlier detection process.}

\item{warn}{\code{logical}. If \strong{FALSE}, warning on whether absolute outliers obtained at a low threshold is indicated. Default \strong{TRUE}.}

\item{verbose}{\code{logical}. Produces messages or not. Default \strong{FALSE}.}

\item{plot}{\code{logical}. to show plot of loess fitted function with minima and maxima (optimal threshold and clean data).}
}
\value{
Either a \code{list} or \code{dataframe} of cleaned records for multiple species.
}
\description{
Optimize threshold for clean data extraction.
}
\examples{

\dontrun{

data(jdsdata)
data(efidata)
matchdata <- match_datasets(datasets = list(jds = jdsdata, efi = efidata),
                            lats = 'lat',
                            lons = 'lon',
                            species = c('speciesname','scientificName'),
                            country= c('JDS4_site_ID'),
                            date=c('sampling_date', 'Date'))

datacheck <- check_names(matchdata, colsp= 'species', pct = 90, merge =TRUE)


db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'), quiet=TRUE)


worldclim <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

rdata <- pred_extract(data = datacheck,
                      raster= worldclim ,
                      lat = 'decimalLatitude',
                      lon= 'decimalLongitude',
                      colsp = 'speciescheck',
                      bbox = db,
                      multiple = TRUE,
                      minpts = 10,
                      list=TRUE,
                      merge=F)


out_df <- multidetect(data = rdata, multiple = TRUE,
                      var = 'bio6',
                      output = 'outlier',
                      exclude = c('x','y'),
                      methods = c('zscore', 'adjbox','iqr', 'semiqr','hampel'))

#extracting optimal threshold for each species

threshopt <- optimal_threshold(refdata = rdata, outliers = out_df)
}



}
