% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plantdf1}
\alias{plantdf1}
\title{Plant species occurences obtained from GBIF, iNaturalist and VertNet.}
\format{
A \code{tibble} 1600 rows and 8 columns.
}
\usage{
data(plantdf1)
}
\description{
A \code{tibble}
}
\details{
The data was incorporated for testing the package only and the licence for citation and further use are included in the file.
}
\examples{

data("plantdf1")
plantdf1

}
\references{
gbif, iNaturalist and VertNet
}
\keyword{Danube}
\keyword{data}
\keyword{plants}
