% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm.R
\name{sdmfit}
\alias{sdmfit}
\title{Species distribution modelling}
\usage{
sdmfit(data, models, cutoff = 0.5, metrics = "all", full = FALSE)
}
\arguments{
\item{data}{Species data list with both testing and training, probably
from boots function.}

\item{models}{The models to be used to examine the relationship between
species occurrences and environmental parameters. Only Random forest
and generalized linear
models are accepted. \code{GLM} is used to set to Generalized Linear Models
and \code{RF} for Random Forest. \code{RF1} variant is slower and is suggested
if RF fails.}

\item{cutoff}{Defines a threshold classify the model predictions/probabilities
as presence or absent. Default is 0.5. The maximum is 1 and lowest is 0.}

\item{metrics}{Either to only consider threshold-dependent \code{(dep)}, threshold-independent \code{(indep)}
or \code{all} evaluation metrics to determine model performance.}

\item{full}{Either full model output for \code{TRUE} or \code{FALSE} for only performance metrics output.}
}
\value{
model runs, and evaluation metrics.
}
\description{
Species distribution modelling
}
\examples{

\dontrun{


data(efidata)

 datacheckf <- subset(efidata,  subset = scientificName \%in\%c("Squalius cephalus"))


#select species with enough records

dataprep <- envextract(occurences = datacheckf, raster = worldclim,
                      lat = "decimalLatitude", lon = "decimalLongitude", binary = FALSE, prop = 0.8)

nb <- boots(data = dataprep, nboots = 10, testprob = 0.3)

sdmd <- sdmfit(data = nb, models = 'RF')
}

}
