% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariatemethods.R
\name{semiIQR}
\alias{semiIQR}
\title{Computes semi-interquantile range to flag suspicious outliers}
\usage{
semiIQR(data, var, output, x = 3)
}
\arguments{
\item{data}{Dataframe to check for outliers}

\item{var}{Environmental parameter considered in flagging suspicious outliers}

\item{output}{Either clean: for dataframe with no suspicious outliers or outlier: to retrun dataframe with only outliers}

\item{x}{A constant to create a fence or boundary to detect outliers.}
}
\value{
Dataframe with or with no outliers.
}
\description{
Computes semi-interquantile range to flag suspicious outliers
}
\examples{

\dontrun{

data("efidata")

gbd <- check_names(data = efidata, colsp='scientificName', pct=90, merge=TRUE)

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'))

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

refdata <- pred_extract(data = gbd, raster= wcd , lat = 'decimalLatitude', lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = db,
                          multiple = FALSE,
                          minpts = 10)

 semiout <- semiIQR(data = refdata[['Salmo trutta']], var = 'bio6', output='outlier')

}



}
