% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliermethods.R
\name{seqfences}
\alias{seqfences}
\title{Title}
\usage{
seqfences(data, var, output, gamma = 0.95, mode = "eo")
}
\arguments{
\item{data}{ll}

\item{var}{lll}

\item{output}{lll}

\item{gamma}{lll}

\item{mode}{lll}
}
\value{
Dataframe or vector with or without outliers
}
\description{
Title
}
\examples{

\dontrun{

data("efidata")

gbd <- check_names(data = efidata, colsp='scientificName', pct=90, merge=TRUE)

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'))

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'), quiet = TRUE)

refdata <- pred_extract(data = gbd, raster= wcd , lat = 'decimalLatitude', lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = db,
                          multiple = FALSE,
                          minpts = 10)

 sqout <- seqfences(data = refdata[['Salmo trutta']], var = 'bio6', output='outlier')
}


}
\references{
\enumerate{

\item Schwertman NC, de Silva R. 2007. Identifying outliers with sequential fences.
Computational Statistics and Data Analysis 51:3800–3810.
\item Schwertman NC, Owens MA, Adnan R. 2004. A simple more general boxplot method for identifying outliers.
Computational Statistics and Data Analysis 47:165–174.

}
}
