% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{suggested.packages}
\alias{suggested.packages}
\title{Check for suggested packages.}
\usage{
suggested.packages(
  listpkgs = c("shiny", "shinydashboard", "DT", "dplyr"),
  reason = "open the R Shiny Application",
  quiet = TRUE
)
}
\arguments{
\item{listpkgs}{Packages to be included.}

\item{reason}{Reason why packages needs to be installed.}

\item{quiet}{Default TRUE, for no messages.}
}
\value{
Packages
}
\description{
Check for suggested packages.
}
