% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesecology.R
\name{thermal_ranges}
\alias{thermal_ranges}
\title{Collates minimum, maximum, and preferable temperatures from FishBase.}
\usage{
thermal_ranges(x, colsp = NULL, verbose = F, pct = 90, sn = FALSE)
}
\arguments{
\item{x}{kk}

\item{colsp}{kk}

\item{verbose}{kk}

\item{pct}{kk}

\item{sn}{kk}
}
\value{
Data table for minimum, maximum and preferable species temperatures from FishBase.
}
\description{
Collates minimum, maximum, and preferable temperatures from FishBase.
}
