% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitybasedmethods.R
\name{xglosh}
\alias{xglosh}
\title{Title}
\usage{
xglosh(data, k, output, exclude = NULL, metric = "manhattan", mode = "soft")
}
\arguments{
\item{data}{uuu}

\item{k}{uuu}

\item{output}{uuu}

\item{exclude}{uzz}

\item{metric}{uuu}

\item{mode}{zzz}
}
\description{
Title
}
