% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliermethods.R
\name{xkmeans}
\alias{xkmeans}
\title{Flags outliers using kmeans clustering method}
\usage{
xkmeans(data, k, exclude, output, mode, method = NULL, verbose = FALSE)
}
\arguments{
\item{data}{Dataframe to check for outliers}

\item{k}{The number of clusters to be used for optimization. It should be greater than 1. For many species k should be be greater 10
to ably cater for each species search for optimal k using the different optimization methods in kmethod}

\item{exclude}{Exclude variables that should not be considered in the fitting the one class model, for example x and y columns or
latitude/longitude or any column that the user doesn't want to consider.}

\item{output}{Either clean: for a data set with no outliers or outlier: to output a data frame with outliers.}

\item{mode}{Either robust, if a robust mode is used which uses median instead of mean and median absolute deviation from median.}

\item{method}{The method to be used for the kmeans clustering. Default is \code{silhouette}.
\code{Elbow method} can be used but user input is required, and therefore multiple outlier detection method
is not possible.}

\item{verbose}{To indicate messages and the default is FALSE.}
}
\value{
Dataframe with or with no outliers.
}
\description{
Flags outliers using kmeans clustering method
}
\examples{

\dontrun{

library(terra)

#species data from online databases

gbdata <- df_retronline(data='Gymnocephalus baloni', gbiflim = 100, inatlim = 100, vertlim = 100)

gbfinal <- merge_all(online = gbdata)

gbchecked <- check_names(data = gbfinal, colsp='species', pct=90, merge=TRUE)

#preclean and extract

danube <- system.file('extdata/danube/basinfinal.shp',  package='specleanr')

danubebasin <- sf::st_read(danube)

#Get environmental data

#worldclim_bio <- env_download(var='bio', resolution = 10, basin = danube, folder='worlclimddata')

worldclim <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'), quiet=TRUE)

precleaned <- pre_cleaner(data = gbchecked,
                          raster= worldclim ,
                          colsp = 'speciescheck',
                          basin = danubebasin,
                          multiple = FALSE,
                          maxpts = 10)

#outliers
kmeanoutliers <- xkmeans(data = precleaned,
                         k=6, exclude = c('x','y'),
                         kmethod ='silhouette',
                         output='outlier')

#clean data
kmeanclean <- xkmeans(data = precleaned,
                         k=6, exclude = c('x','y'),
                         kmethod ='silhouette',
                         output='clean')
}


}
