% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{xkmedian}
\alias{xkmedian}
\title{Using kmedian outlier detection methods}
\usage{
xkmedian(data, k = 3)
}
\arguments{
\item{data}{Input data for for checking outliers}

\item{k}{Number of clusters}
}
\value{
clean data set after kmedian outlier cleaning.
}
\description{
Using kmedian outlier detection methods
}
