% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliermethods.R
\name{xkmedian}
\alias{xkmedian}
\title{Using kmedian outlier detection methods}
\usage{
xkmedian(data, k = 3)
}
\arguments{
\item{data}{Input data for for checking outliers}

\item{k}{Number of clusters}
}
\value{
clean or outlier data set after outlier detection.
}
\description{
Using kmedian outlier detection methods
}
\examples{

\dontrun{

data("efidata")

gbd <- check_names(data = efidata, colsp='scientificName', pct=90, merge=TRUE)

db <- sf::st_read(system.file('extdata/danube/basinfinal.shp', package='specleanr'), quiet = TRUE)

wcd <- terra::rast(system.file('extdata/worldclim.tiff', package='specleanr'))

refdata <- pred_extract(data = gbd, raster= wcd ,
                       lat = 'decimalLatitude',
                       lon= 'decimalLongitude',
                       colsp = 'speciescheck',
                      bbox = db,
                       multiple = TRUE,
                       minpts = 10)

kmedianout <- kmedian(data = refdata, k = 3)

}

}
