% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{xkmedoid}
\alias{xkmedoid}
\title{Title}
\usage{
xkmedoid(data, k = 2, metric = "manhattan", output = "clean", exclude, x = 1.5)
}
\arguments{
\item{data}{kk}

\item{k}{kk}

\item{metric}{kk}

\item{output}{kk}

\item{exclude}{kk}

\item{x}{a constant to detemrine outliers.}
}
\description{
Title
}
