% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitybasedmethods.R
\name{xknn}
\alias{xknn}
\title{Title}
\usage{
xknn(data, output, exclude = NULL, metric = "manhattan", mode = "soft")
}
\arguments{
\item{data}{zzz}

\item{output}{zztt}

\item{exclude}{zzuu}

\item{metric}{ttt}

\item{mode}{uiii}
}
\description{
Title
}
