% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliermethods.R
\name{xlof}
\alias{xlof}
\title{Flags suspicious using the local outlier factor or Density-Based Spatial Clustering of Applications with Noise}
\usage{
xlof(data, output, minPts, exclude = NULL, metric = "manhattan", mode = "soft")
}
\arguments{
\item{data}{Data frame of species records with environmental data}

\item{output}{Either clean: for data frame with no suspicious outliers or outlier: to return dataframe with only outliers.}

\item{minPts}{Minimum neighbors around the records.}

\item{exclude}{Exclude variables that should not be considered in the fitting the one class model, for example x and y columns or
latitude/longitude or any column that the user doesn't want to consider.}

\item{metric}{Distance-based measure to examine the distance between variables. Default \code{manhattan}.}

\item{mode}{Either \code{soft} if mean is used or \code{robust} if mad is used. Defualt \code{soft}.}
}
\value{
Dataframe with or with no outliers.
}
\description{
Flags suspicious using the local outlier factor or Density-Based Spatial Clustering of Applications with Noise
}
\examples{

\dontrun{

library(terra)

#species data from online databases

gbdata <- df_retronline(data='Gymnocephalus baloni', gbiflim = 100, inatlim = 100, vertlim = 100)

gbfinal <- merge_all(online = gbdata)

gbchecked <- check_names(data = gbfinal, colsp='species', pct=90, merge=TRUE)

#preclean and extract

danube <- system.file('extdata/danube/basinfinal.shp', quiet=TRUE, package='specleanr')

danubebasin <- sf::st_read(danube)

#Get environmental data

#worldclim_bio <- env_download(var='bio', resolution = 10, basin = danube, folder='worlclimddata')

worldclim <- terra::rast(system.file('extdata/worldclim.tiff', quiet=TRUE, package='specleanr'))

precleaned <- precleaner(data = gbchecked,
                          raster= worldclim ,
                          lat = 'decimalLatitude',
                          lon= 'decimalLongitude',
                          colsp = 'speciescheck',
                          basin = danubebasin,
                          multiple = FALSE,
                          minpts = 10)

#outliers
lof_outliers <- xlof(data = precleaned,
                         exclude = c('x','y'),
                         output='outlier')

#clean data
lof_clean <- xlof(data = precleaned,
                         exclude = c('x','y'),
                         output='clean')



}

}
